% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-discret_method.R
\name{method_func}
\alias{method_func}
\alias{method_func,discretization_method-method}
\alias{method_func<-}
\alias{method_func<-,discretization_method-method}
\alias{spm_name,discretization_method-method}
\alias{spm_name<-,discretization_method-method}
\title{Accessing OR replacing \code{discretization_method} model elements}
\usage{
method_func(sspm_object)

\S4method{method_func}{discretization_method}(sspm_object)

method_func(object) <- value

\S4method{method_func}{discretization_method}(object) <- value

\S4method{spm_name}{discretization_method}(sspm_object)

\S4method{spm_name}{discretization_method}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[discretization_method]} An object of class
\link[=discretization_method-class]{discretization_method}.}

\item{object}{\strong{[discretization_method]} An object of class
\link[=discretization_method-class]{discretization_method}.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
The object in the required slot.
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=discretization_method-class]{discretization_method}.
}
\examples{
method <- as_discretization_method("tesselate_voronoi")
method_func(method)
}
