% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-spaspm_formula.R
\name{raw_formula}
\alias{raw_formula}
\alias{raw_formula,spaspm_formula-method}
\alias{raw_formula<-}
\alias{raw_formula<-,spaspm_formula-method}
\alias{translated_formula,spaspm_formula-method}
\alias{translated_formula<-}
\alias{translated_formula<-,spaspm_formula-method}
\alias{dataset,spaspm_formula-method}
\alias{dataset<-}
\alias{dataset<-,spaspm_formula-method}
\alias{formula_vars,spaspm_formula-method}
\alias{formula_vars<-}
\alias{formula_vars<-,spaspm_formula-method}
\title{Accessing OR replacing \code{spaspm_formula} model elements}
\usage{
raw_formula(spaspm_object)

\S4method{raw_formula}{spaspm_formula}(spaspm_object)

raw_formula(object) <- value

\S4method{raw_formula}{spaspm_formula}(object) <- value

\S4method{translated_formula}{spaspm_formula}(spaspm_object)

translated_formula(object) <- value

\S4method{translated_formula}{spaspm_formula}(object) <- value

\S4method{dataset}{spaspm_formula}(spaspm_object)

dataset(object) <- value

\S4method{dataset}{spaspm_formula}(object) <- value

\S4method{formula_vars}{spaspm_formula}(spaspm_object)

formula_vars(object) <- value

\S4method{formula_vars}{spaspm_formula}(object) <- value
}
\arguments{
\item{spaspm_object}{\strong{[spaspm_formula]} An object of class
\link[=spaspm_formula-class]{spaspm_formula}.}

\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=spaspm_formula-class]{spaspm_formula}.
}
\section{Functions}{
\itemize{
\item \code{raw_formula,spaspm_formula-method}: TODO

\item \code{raw_formula<-}: TODO

\item \code{raw_formula<-,spaspm_formula-method}: TODO

\item \code{translated_formula,spaspm_formula-method}: TODO

\item \code{translated_formula<-}: TODO

\item \code{translated_formula<-,spaspm_formula-method}: TODO

\item \code{dataset,spaspm_formula-method}: TODO

\item \code{dataset<-}: TODO

\item \code{dataset<-,spaspm_formula-method}: TODO

\item \code{formula_vars,spaspm_formula-method}: TODO

\item \code{formula_vars<-}: TODO

\item \code{formula_vars<-,spaspm_formula-method}: TODO
}}

