% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-sspm_boundary.R
\name{spm_boundaries,sspm_boundary-method}
\alias{spm_boundaries,sspm_boundary-method}
\alias{spm_boundaries<-,sspm_boundary-method}
\alias{spm_discret_method}
\alias{spm_discret_method,sspm_boundary-method}
\alias{spm_discret_method<-}
\alias{spm_discret_method<-,sspm_boundary-method}
\alias{spm_patches}
\alias{spm_patches,sspm_boundary-method}
\alias{spm_patches<-}
\alias{spm_patches<-,sspm_boundary-method}
\alias{spm_points}
\alias{spm_points,sspm_boundary-method}
\alias{spm_points<-}
\alias{spm_points<-,sspm_boundary-method}
\alias{spm_boundary_colum}
\alias{spm_boundary_colum,sspm_boundary-method}
\alias{spm_boundary_colum<-}
\alias{spm_boundary_colum<-,sspm_boundary-method}
\alias{spm_surface_colum}
\alias{spm_surface_colum,sspm_boundary-method}
\alias{spm_surface_colum<-}
\alias{spm_surface_colum<-,sspm_boundary-method}
\title{Accessing OR replacing \code{sspm_boundary} model elements}
\usage{
\S4method{spm_boundaries}{sspm_boundary}(sspm_object)

\S4method{spm_boundaries}{sspm_boundary}(object) <- value

spm_discret_method(sspm_object)

\S4method{spm_discret_method}{sspm_boundary}(sspm_object)

spm_discret_method(object) <- value

\S4method{spm_discret_method}{sspm_boundary}(object) <- value

spm_patches(sspm_object)

\S4method{spm_patches}{sspm_boundary}(sspm_object)

spm_patches(object) <- value

\S4method{spm_patches}{sspm_boundary}(object) <- value

spm_points(sspm_object)

\S4method{spm_points}{sspm_boundary}(sspm_object)

spm_points(object) <- value

\S4method{spm_points}{sspm_boundary}(object) <- value

spm_boundary_colum(sspm_object)

\S4method{spm_boundary_colum}{sspm_boundary}(sspm_object)

spm_boundary_colum(object) <- value

\S4method{spm_boundary_colum}{sspm_boundary}(object) <- value

spm_surface_colum(sspm_object)

\S4method{spm_surface_colum}{sspm_boundary}(sspm_object)

spm_surface_colum(object) <- value

\S4method{spm_surface_colum}{sspm_boundary}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm_boundary]} An object of class
\link[=sspm_boundary-class]{sspm_boundary}.}

\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm_boundary-class]{sspm_boundary}.
}
