% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-sspm_data.R
\name{spm_data}
\alias{spm_data}
\alias{spm_data,sspm-method}
\alias{spm_data,sspm_data-method}
\alias{spm_data<-}
\alias{spm_data<-,sspm_data-method}
\alias{spm_name,sspm_data-method}
\alias{spm_name<-,sspm_data-method}
\alias{spm_unique_ID}
\alias{spm_unique_ID,sspm_data-method}
\alias{spm_unique_ID<-}
\alias{spm_unique_ID<-,sspm_data-method}
\alias{spm_coords_col}
\alias{spm_coords_col,sspm_data-method}
\alias{spm_coords_col<-}
\alias{spm_coords_col<-,sspm_data-method}
\alias{spm_time_column}
\alias{spm_time_column,sspm_data-method}
\alias{spm_time_column<-}
\alias{spm_time_column<-,sspm_data-method}
\alias{spm_formulas,sspm_data-method}
\alias{spm_formulas<-,sspm_data-method}
\alias{spm_type}
\alias{spm_type,sspm_data-method}
\alias{spm_type<-}
\alias{spm_type<-,sspm_data-method}
\alias{spm_smoothed_data,sspm_data-method}
\alias{spm_smoothed_data<-,sspm_data-method}
\alias{spm_smoothed_fit}
\alias{spm_smoothed_fit,sspm_data-method}
\alias{spm_smoothed_fit<-}
\alias{spm_smoothed_fit<-,sspm_data-method}
\alias{is_smoothed}
\alias{is_smoothed,sspm_data-method}
\alias{is_smoothed<-}
\alias{is_smoothed<-,sspm_data-method}
\alias{is_splitted}
\alias{is_splitted,sspm_data-method}
\alias{is_splitted<-}
\alias{is_splitted<-,sspm_data-method}
\title{Accessing OR replacing \code{sspm_data} model elements}
\usage{
spm_data(sspm_object)

\S4method{spm_data}{sspm}(sspm_object)

\S4method{spm_data}{sspm_data}(sspm_object)

spm_data(object) <- value

\S4method{spm_data}{sspm_data}(object) <- value

\S4method{spm_name}{sspm_data}(sspm_object)

\S4method{spm_name}{sspm_data}(object) <- value

spm_unique_ID(sspm_object)

\S4method{spm_unique_ID}{sspm_data}(sspm_object)

spm_unique_ID(object) <- value

\S4method{spm_unique_ID}{sspm_data}(object) <- value

spm_coords_col(sspm_object)

\S4method{spm_coords_col}{sspm_data}(sspm_object)

spm_coords_col(object) <- value

\S4method{spm_coords_col}{sspm_data}(object) <- value

spm_time_column(sspm_object)

\S4method{spm_time_column}{sspm_data}(sspm_object)

spm_time_column(object) <- value

\S4method{spm_time_column}{sspm_data}(object) <- value

\S4method{spm_formulas}{sspm_data}(sspm_object)

\S4method{spm_formulas}{sspm_data}(object) <- value

spm_type(sspm_object)

\S4method{spm_type}{sspm_data}(sspm_object)

spm_type(object) <- value

\S4method{spm_type}{sspm_data}(object) <- value

\S4method{spm_smoothed_data}{sspm_data}(sspm_object)

\S4method{spm_smoothed_data}{sspm_data}(object) <- value

spm_smoothed_fit(sspm_object)

\S4method{spm_smoothed_fit}{sspm_data}(sspm_object)

spm_smoothed_fit(object) <- value

\S4method{spm_smoothed_fit}{sspm_data}(object) <- value

is_smoothed(sspm_object)

\S4method{is_smoothed}{sspm_data}(sspm_object)

is_smoothed(object) <- value

\S4method{is_smoothed}{sspm_data}(object) <- value

is_splitted(sspm_object)

\S4method{is_splitted}{sspm_data}(sspm_object)

is_splitted(object) <- value

\S4method{is_splitted}{sspm_data}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm_data]} An object of class
\link[=sspm_data-class]{sspm_data}.}

\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm_data-class]{sspm_data}.
}
