% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-sspm_dataset.R
\name{spm_data}
\alias{spm_data}
\alias{spm_data,sspm_dataset-method}
\alias{spm_data<-}
\alias{spm_data<-,sspm_dataset-method}
\alias{spm_name,sspm_dataset-method}
\alias{spm_name<-,sspm_dataset-method}
\alias{spm_unique_ID,sspm_dataset-method}
\alias{spm_unique_ID<-,sspm_dataset-method}
\alias{spm_coords_col}
\alias{spm_coords_col,sspm_dataset-method}
\alias{spm_coords_col<-}
\alias{spm_coords_col<-,sspm_dataset-method}
\alias{spm_time,sspm_dataset-method}
\alias{spm_time<-,sspm_dataset-method}
\alias{spm_biomass_vars}
\alias{spm_biomass_vars,sspm_dataset-method}
\alias{spm_biomass_vars<-}
\alias{spm_biomass_vars<-,sspm_dataset-method}
\alias{spm_density_vars}
\alias{spm_density_vars,sspm_dataset-method}
\alias{spm_density_vars<-}
\alias{spm_density_vars<-,sspm_dataset-method}
\alias{spm_formulas}
\alias{spm_formulas,sspm_dataset-method}
\alias{spm_formulas<-}
\alias{spm_formulas<-,sspm_dataset-method}
\alias{spm_smoothed_data,sspm_dataset-method}
\alias{spm_smoothed_data<-,sspm_dataset-method}
\alias{spm_smoothed_fit}
\alias{spm_smoothed_fit,sspm_dataset-method}
\alias{spm_smoothed_fit<-}
\alias{spm_smoothed_fit<-,sspm_dataset-method}
\alias{spm_smoothed_vars}
\alias{spm_smoothed_vars,sspm_dataset-method}
\alias{spm_smoothed_vars<-}
\alias{spm_smoothed_vars<-,sspm_dataset-method}
\alias{is_mapped}
\alias{is_mapped,sspm_dataset-method}
\alias{is_mapped<-}
\alias{is_mapped<-,sspm_dataset-method}
\alias{spm_boundaries,sspm_dataset-method}
\alias{spm_boundaries<-,sspm_dataset-method}
\title{Accessing OR replacing \code{sspm_dataset} model elements}
\usage{
spm_data(sspm_object)

\S4method{spm_data}{sspm_dataset}(sspm_object)

spm_data(object) <- value

\S4method{spm_data}{sspm_dataset}(object) <- value

\S4method{spm_name}{sspm_dataset}(sspm_object)

\S4method{spm_name}{sspm_dataset}(object) <- value

\S4method{spm_unique_ID}{sspm_dataset}(sspm_object)

\S4method{spm_unique_ID}{sspm_dataset}(object) <- value

spm_coords_col(sspm_object)

\S4method{spm_coords_col}{sspm_dataset}(sspm_object)

spm_coords_col(object) <- value

\S4method{spm_coords_col}{sspm_dataset}(object) <- value

\S4method{spm_time}{sspm_dataset}(sspm_object)

\S4method{spm_time}{sspm_dataset}(object) <- value

spm_biomass_vars(sspm_object)

\S4method{spm_biomass_vars}{sspm_dataset}(sspm_object)

spm_biomass_vars(object) <- value

\S4method{spm_biomass_vars}{sspm_dataset}(object) <- value

spm_density_vars(sspm_object)

\S4method{spm_density_vars}{sspm_dataset}(sspm_object)

spm_density_vars(object) <- value

\S4method{spm_density_vars}{sspm_dataset}(object) <- value

spm_formulas(sspm_object)

\S4method{spm_formulas}{sspm_dataset}(sspm_object)

spm_formulas(object) <- value

\S4method{spm_formulas}{sspm_dataset}(object) <- value

\S4method{spm_smoothed_data}{sspm_dataset}(sspm_object)

\S4method{spm_smoothed_data}{sspm_dataset}(object) <- value

spm_smoothed_fit(sspm_object)

\S4method{spm_smoothed_fit}{sspm_dataset}(sspm_object)

spm_smoothed_fit(object) <- value

\S4method{spm_smoothed_fit}{sspm_dataset}(object) <- value

spm_smoothed_vars(sspm_object)

\S4method{spm_smoothed_vars}{sspm_dataset}(sspm_object)

spm_smoothed_vars(object) <- value

\S4method{spm_smoothed_vars}{sspm_dataset}(object) <- value

is_mapped(sspm_object)

\S4method{is_mapped}{sspm_dataset}(sspm_object)

is_mapped(object) <- value

\S4method{is_mapped}{sspm_dataset}(object) <- value

\S4method{spm_boundaries}{sspm_dataset}(sspm_object)

\S4method{spm_boundaries}{sspm_dataset}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm_dataset]} An object of class
\link[=sspm_dataset-class]{sspm_dataset}.}

\item{object}{\strong{[sspm_dataset]} An object of class
\link[=sspm_dataset-class]{sspm_dataset}.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
The object in the required slot.
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm_dataset-class]{sspm_dataset}.
}
\examples{
data(borealis_simulated, package = "sspm")
biomass_dataset <- spm_as_dataset(data.frame(borealis_simulated), name = "borealis",
                                  density = "weight_per_km2",
                                  time = "year_f",
                                  coords = c('lon_dec','lat_dec'),
                                  uniqueID = "uniqueID")
spm_data(biomass_dataset)
}
