% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-sspm_formula.R
\name{raw_formula}
\alias{raw_formula}
\alias{raw_formula,sspm_formula-method}
\alias{raw_formula<-}
\alias{raw_formula<-,sspm_formula-method}
\alias{translated_formula}
\alias{translated_formula,sspm_formula-method}
\alias{translated_formula<-}
\alias{translated_formula<-,sspm_formula-method}
\alias{formula_vars}
\alias{formula_vars,sspm_formula-method}
\alias{formula_vars<-}
\alias{formula_vars<-,sspm_formula-method}
\alias{formula_type}
\alias{formula_type,sspm_formula-method}
\alias{formula_type<-}
\alias{formula_type<-,sspm_formula-method}
\alias{is_fitted}
\alias{is_fitted,sspm_formula-method}
\alias{is_fitted<-}
\alias{is_fitted<-,sspm_formula-method}
\alias{spm_response}
\alias{spm_response,sspm_formula-method}
\alias{spm_response<-}
\alias{spm_response<-,sspm_formula-method}
\alias{spm_lagged_vars}
\alias{spm_lagged_vars,sspm_formula-method}
\alias{spm_lagged_vars<-}
\alias{spm_lagged_vars<-,sspm_formula-method}
\title{Accessing OR replacing \code{sspm_formula} model elements}
\usage{
raw_formula(sspm_object)

\S4method{raw_formula}{sspm_formula}(sspm_object)

raw_formula(object) <- value

\S4method{raw_formula}{sspm_formula}(object) <- value

translated_formula(sspm_object)

\S4method{translated_formula}{sspm_formula}(sspm_object)

translated_formula(object) <- value

\S4method{translated_formula}{sspm_formula}(object) <- value

formula_vars(sspm_object)

\S4method{formula_vars}{sspm_formula}(sspm_object)

formula_vars(object) <- value

\S4method{formula_vars}{sspm_formula}(object) <- value

formula_type(sspm_object)

\S4method{formula_type}{sspm_formula}(sspm_object)

formula_type(object) <- value

\S4method{formula_type}{sspm_formula}(object) <- value

is_fitted(sspm_object)

\S4method{is_fitted}{sspm_formula}(sspm_object)

is_fitted(object) <- value

\S4method{is_fitted}{sspm_formula}(object) <- value

spm_response(sspm_object)

\S4method{spm_response}{sspm_formula}(sspm_object)

spm_response(object) <- value

\S4method{spm_response}{sspm_formula}(object) <- value

spm_lagged_vars(sspm_object)

\S4method{spm_lagged_vars}{sspm_formula}(sspm_object)

spm_lagged_vars(object) <- value

\S4method{spm_lagged_vars}{sspm_formula}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm_formula]} An object of class
\link[=sspm_formula-class]{sspm_formula}.}

\item{object}{\strong{[sspm_formula]} An object of class
\link[=sspm_formula-class]{sspm_formula}.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
The object in the required slot.
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm_formula-class]{sspm_formula}.
}
\examples{
form <- new("sspm_formula",
            raw_formula = as.formula("weight_per_km2 ~ smooth_time()"),
            translated_formula = as.formula("weight_per_km2 ~ s(year_f,
                      k = 24L, bs = 're', xt = list(penalty = pen_mat_time))"),
                    vars = list(pen_mat_time = matrix(),
                                pen_mat_space = matrix()),
                    response = "weight_per_km2")
translated_formula(form)
}
