% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_discretization_method.R
\name{as_discretization_method}
\alias{as_discretization_method}
\alias{as_discretization_method,character-method}
\title{Cast into a discretization_method object}
\usage{
as_discretization_method(method, ...)

\S4method{as_discretization_method}{character}(method, ...)
}
\description{
TODO
}
\section{Functions}{
\itemize{
\item \code{as_discretization_method,character-method}: TODO
}}

