% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_spaspm_data.R
\name{as_spaspm_data}
\alias{as_spaspm_data}
\alias{as_spaspm_data,spaspm_data,ANY-method}
\alias{as_spaspm_data,data.frame,NULL-method}
\alias{as_spaspm_data,data.frame,character-method}
\alias{as_spaspm_data,sf,ANY-method}
\title{Create a \code{spaspm_data} dataset structure}
\usage{
as_spaspm_data(data, coords, name, uniqueID, crs, ...)

\S4method{as_spaspm_data}{spaspm_data,ANY}(data, coords, name, uniqueID, crs, ...)

\S4method{as_spaspm_data}{data.frame,`NULL`}(data, coords, name, uniqueID, crs, ...)

\S4method{as_spaspm_data}{data.frame,character}(data, coords, name, uniqueID, crs, ...)

\S4method{as_spaspm_data}{sf,ANY}(data, coords, name, uniqueID, crs, ...)
}
\arguments{
\item{data}{\strong{[data.frame OR sf]} The dataset.}

\item{coords}{\strong{[character]} The column of \code{data} for longitude and
latitude of the observations.}

\item{name}{\strong{[character]} The name of the dataset, default to "Biomass".}

\item{uniqueID}{\strong{[character]} The column of \code{data} that is unique for all
rows of the data matrix.}

\item{crs}{Coordinate reference system, passed onto \link[=sf]{st_as_sf}.}
}
\description{
Create a \code{spaspm_data} dataset structure
}
\section{Functions}{
\itemize{
\item \code{as_spaspm_data,spaspm_data,ANY-method}: TODO

\item \code{as_spaspm_data,data.frame,NULL-method}: TODO

\item \code{as_spaspm_data,data.frame,character-method}: TODO

\item \code{as_spaspm_data,sf,ANY-method}: TODO
}}

