% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_spaspm_smooth.R
\name{as_spaspm_smooth}
\alias{as_spaspm_smooth}
\alias{as_spaspm_smooth,spaspm_smooth-method}
\title{Create a \code{spaspm_smooth} object}
\usage{
as_spaspm_smooth(smooth, type, args)

\S4method{as_spaspm_smooth}{spaspm_smooth}(smooth)
}
\arguments{
\item{smooth}{\strong{[mgcv xxx.smooth.spec object]} A smooth object from \link{mgcv}.}

\item{type}{\strong{[character]}}

\item{args}{\strong{[list]}}

\item{dataset_name}{\strong{[character]} The name of the dataset on which to
map the smooth. See also \link[=spaspm_smooth-class]{spaspm_smooth}.}
}
\value{
An object of class \code{\link[=spaspm_smooth-class]{spaspm_smooth}}.
}
\description{
This function allows to cast different inputs into an object of class
\code{\link[=spaspm_smooth-class]{spaspm_smooth}}.
}
\section{Functions}{
\itemize{
\item \code{as_spaspm_smooth,spaspm_smooth-method}: TODO
}}

