% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_simulated}
\alias{catch_simulated}
\title{Simulated catch data}
\format{
A data frame:
\describe{
\item{year_f}{Year as a factor}
\item{sfa}{SFA ID number}
\item{catch}{Simualated catch in kg}
\item{lon_dec}{Longitude}
\item{lat_dec}{Latitude}
\item{row}{Row ID}
\item{uniqueID}{Unique ID for simulated observation}
}
}
\usage{
catch_simulated
}
\description{
Simulated catch data for test and practice.
}
\keyword{datasets}
