% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_S4_all_classes_definitions.R
\docType{class}
\name{discretization_method-class}
\alias{discretization_method-class}
\title{sspm discretization method class}
\description{
This class encapsulates a name and a method (function) used for
discretization.
}
\section{Slots}{

\describe{
\item{\code{name}}{\strong{[character]} Name of the discretization method.}

\item{\code{method}}{\strong{[function]} Function used for discretization.}
}}

