% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_formula.R
\name{map_formula}
\alias{map_formula}
\alias{map_formula,spaspm,ANY,ANY-method}
\alias{map_formula,spaspm_discrete,missing,ANY-method}
\alias{map_formula,spaspm_discrete,ANY,missing-method}
\alias{map_formula,spaspm_discrete,character,formula-method}
\title{Map model formula onto a discretized SPASPM object}
\usage{
map_formula(spaspm_object, dataset, formula, ...)

\S4method{map_formula}{spaspm,ANY,ANY}(spaspm_object, dataset, formula, ...)

\S4method{map_formula}{spaspm_discrete,missing,ANY}(spaspm_object, dataset, formula, ...)

\S4method{map_formula}{spaspm_discrete,ANY,missing}(spaspm_object, dataset, formula, ...)

\S4method{map_formula}{spaspm_discrete,character,formula}(spaspm_object, dataset, formula, ...)
}
\arguments{
\item{spaspm_object}{\strong{[spaspm_discrete]} An object of class
\link[=spaspm_discrete-class]{spaspm_discrete}.}

\item{dataset}{\strong{[character]} The name of the dataset among base and/or
mapped datasets for which to specify the formula}

\item{formula}{\strong{[formula]} A formula definition of the form
response ~ smoothing_terms + ...}

\item{...}{Further arguments passed down, none used for now.}
}
\description{
This functions allows to specify a model formula for a given discrete SPASPM
object. This formula makes use of specific smoothing terms \code{smooth_time()},
\code{smooth_space()}, \code{smooth_space_time()}. This formula can also contain fixed
effects and custom smooths, and can make use of specific smoothing terms
\code{smooth_time()}, \code{smooth_space()}, \code{smooth_space_time()}. See Details for
more explanations.
}
\section{Functions}{
\itemize{
\item \code{map_formula,spaspm,ANY,ANY-method}: TODO

\item \code{map_formula,spaspm_discrete,missing,ANY-method}: TODO

\item \code{map_formula,spaspm_discrete,ANY,missing-method}: TODO

\item \code{map_formula,spaspm_discrete,character,formula-method}: TODO
}}

