% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_smooth.R
\name{map_smooth}
\alias{map_smooth}
\alias{map_smooth,spaspm-method}
\alias{map_smooth,spaspm_discrete-method}
\title{Specify a smoothing model for a discrete \code{spaspm} object}
\usage{
map_smooth(
  spaspm_object,
  dataset_name = "Biomass",
  time = NULL,
  space = NULL,
  space_time = NULL,
  ...
)

\S4method{map_smooth}{spaspm}(
  spaspm_object,
  dataset_name = "Biomass",
  time = NULL,
  space = NULL,
  space_time = NULL,
  ...
)

\S4method{map_smooth}{spaspm_discrete}(
  spaspm_object,
  dataset_name = "Biomass",
  time = NULL,
  space = NULL,
  space_time = NULL,
  ...
)
}
\arguments{
\item{spaspm_object}{\strong{[spaspm_discrete]} An object of class
\link[=spaspm_discrete-class]{spaspm_discrete}.}

\item{dataset_name}{\strong{[character]} The name of the dataset of the
object to map the smooth for.}

\item{time}{\strong{[list]} Specifications for the modeling the temporal
dimension of the smooth model.}

\item{space}{\strong{[list]} Specifications for the modeling the spatial
dimension of the smooth model.}

\item{space_time}{\strong{[list]} Specifications for the modeling the
spatio-temporal dimension of the smooth model.}

\item{...}{Other arguments, none used at the moment.}
}
\value{
The updated object, of class \link[=spaspm_discrete-class]{spaspm_discrete}.
}
\description{
Allows to specify a smoothing model for a discrete \code{spaspm} object.
}
\section{Functions}{
\itemize{
\item \code{map_smooth,spaspm-method}: TODO

\item \code{map_smooth,spaspm_discrete-method}: TODO
}}

