% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm-plot.R
\name{plot.spaspm}
\alias{plot.spaspm}
\alias{plot,spaspm-method}
\alias{plot,spaspm_discrete-method}
\title{Plot a discrete \code{spaspm} model object}
\usage{
\S4method{plot}{spaspm}(x, y)

\S4method{plot}{spaspm_discrete}(x, y)
}
\arguments{
\item{x}{\strong{[spaspm_discrete]} An object of class
\link[=spaspm_discrete-class]{spaspm_discrete}.}
}
\value{
N/A
}
\description{
This will plot the patches and points of a \code{spaspm_discrete} object.
}
\section{Functions}{
\itemize{
\item \code{plot,spaspm-method}: TODO

\item \code{plot,spaspm_discrete-method}: TODO
}}

