% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspm-plot.R
\name{plot.sspm}
\alias{plot.sspm}
\alias{plot,sspm_boundary-method}
\alias{plot,sspm_discrete-method}
\title{Plot various \code{sspm} objects}
\usage{
\S4method{plot}{sspm_boundary}(x)

\S4method{plot}{sspm_discrete}(x, smoothed_var = NULL, page = "first", nrow = 2, ncol = 4)
}
\arguments{
\item{x}{An object from this package.}

\item{smoothed_var}{TODO}

\item{page}{TODO}

\item{nrow}{TODO}

\item{ncol}{TODO}
}
\value{
N/A
}
\description{
Plot methods for a range of sspm objects.
}
