% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.sspm}
\alias{plot,sspm_boundary,missing-method}
\alias{plot,sspm_dataset,missing-method}
\alias{plot,sspm_fit,missing-method}
\title{Plot \code{sspm} objects}
\usage{
\S4method{plot}{sspm_boundary,missing}(x, y, ...)

\S4method{plot}{sspm_dataset,missing}(
  x,
  y,
  ...,
  var = NULL,
  use_sf = FALSE,
  page = "first",
  nrow = 2,
  ncol = 4,
  log = FALSE,
  scales = "fixed"
)

\S4method{plot}{sspm_fit,missing}(
  x,
  y,
  ...,
  train_test = FALSE,
  biomass = FALSE,
  var = NULL,
  biomass_var_predict = NULL,
  biomass_var_smooth = NULL,
  biomass_var_origin = NULL,
  use_sf = FALSE,
  page = "first",
  nrow = 2,
  ncol = 4,
  log = FALSE,
  scales = "fixed"
)
}
\arguments{
\item{x}{\strong{[sspm_...]} An object from this package.}

\item{y}{NOT USED (from generic).}

\item{...}{NOT USED (from generic).}

\item{var}{\strong{[character]} (for sspm_dataset) Variable to plot.}

\item{use_sf}{\strong{[logical]} Whether to produce a spatial plot.}

\item{page}{The page to draw}

\item{nrow}{Number of rows per page}

\item{ncol}{Number of columns per page}

\item{log}{\strong{[logical]} Whether to plot on a log scale, default to TRUE.}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{train_test}{\strong{[logical]} (For sspm_fit) Whether to plot a train/test
pair plot.}

\item{biomass}{\strong{[logical]} (For sspm_fit) Whether to plot a biomass
pair plot.}

\item{biomass_var_predict}{\strong{[character]} Biomass variable to plot (from
predictions, OPTIONNAL).}

\item{biomass_var_smooth}{\strong{[character]} Biomass variable to plot (smooth,
optional).}

\item{biomass_var_origin}{\strong{[character]} Biomass variable to plot (from
original dataset, optionnal)}
}
\value{
A ggplot object.
}
\description{
Plot methods for a range of sspm objects.
}
