% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predator_simulated}
\alias{predator_simulated}
\title{Simulated predator data}
\format{
A data frame:
\describe{
\item{year_f}{Year as a factor}
\item{sfa}{SFA ID number}
\item{weight_per_km2}{Simualated biomass in kg per km2}
\item{lon_dec}{Longitude}
\item{lat_dec}{Latitude}
\item{row}{Row ID}
\item{uniqueID}{Unique ID for simulated observation}
}
}
\usage{
predator_simulated
}
\description{
Simulated predator data for test and practice.
}
\keyword{datasets}
