% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.sspm}
\alias{predict,sspm_fit-method}
\title{Predict with a SPM model}
\usage{
\S4method{predict}{sspm_fit}(object, new_data = NULL, biomass = NULL, next_ts = FALSE, ...)
}
\arguments{
\item{object}{\emph{[sspm_fit]}* Fit object to predict from.}

\item{new_data}{\strong{[data.frame]} New data to predict with.}

\item{biomass}{\strong{[character]} Biomass variable.}

\item{next_ts}{\strong{[logical]} For biomass, predict next timestep.}

\item{...}{Arguments passed on to \link{predict.bam}.}
}
\value{
A \code{dataframe} of predictions.
}
\description{
Predict using a fitted SPM model on the whole data or on new data
}
