% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_boundaries.R
\docType{data}
\name{sfa_boundaries}
\alias{sfa_boundaries}
\title{SFA boundaries data}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{sfa}{SFA ID number}
\item{geometry}{sf geometry}
}
}
\source{
\url{https://www.dfo-mpo.gc.ca/fisheries-peches/ifmp-gmp/shrimp-crevette/shrimp-crevette-2018-002-eng.html}
}
\usage{
sfa_boundaries
}
\description{
SFA boundaries.
}
\keyword{datasets}
