% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooths.R
\name{smooth_time}
\alias{smooth_time}
\alias{smooth_space}
\alias{smooth_space_time}
\title{SPASPM Smoothing functions}
\usage{
smooth_time(type)

smooth_space(type)

smooth_space_time(type)
}
\arguments{
\item{type}{\strong{[character]} Type of smooths.}
}
\description{
A full spaspm formula contains calls to the smoothing terms \code{smooth_time()},
\code{smooth_space()}, \code{smooth_space_time()}.
}
\section{Functions}{
\itemize{
\item \code{smooth_space}: TODO

\item \code{smooth_space_time}: TODO
}}

