% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm-accessors-methods.R
\name{spm_name}
\alias{spm_name}
\title{Accessing OR replacing \code{spaspm} model elements}
\usage{
spm_name(spaspm_object)
}
\arguments{
\item{spaspm_object}{\strong{[spaspm OR adjacent]} An object of class
\link[=spaspm-class]{spaspm} or others derivative classes.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=spaspm-class]{spaspm} and others derivative classes
(\code{spaspm_discrete}, etc...).
}
