% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm.R
\name{spaspm}
\alias{spaspm}
\title{Create a spaspm model object}
\usage{
spaspm(
  model_name = "My SPASPM model",
  dataset_name = "Biomass",
  data,
  uniqueID,
  boundaries,
  coords = NULL,
  ...
)
}
\description{
TODO
}
