% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm_data-accessors-methods.R
\name{spm_data}
\alias{spm_data}
\alias{spm_data,spaspm-method}
\alias{spm_data,spaspm_data-method}
\title{Accessing OR replacing \code{spaspm_data} model elements}
\usage{
spm_data(spaspm_object)

\S4method{spm_data}{spaspm}(spaspm_object)

\S4method{spm_data}{spaspm_data}(spaspm_object)
}
\arguments{
\item{spaspm_object}{\strong{[spaspm_data]} An object of class
\link[=spaspm_data-class]{spaspm_data}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=spaspm_data-class]{spaspm_data}.
}
\section{Functions}{
\itemize{
\item \code{spm_data,spaspm-method}: TODO

\item \code{spm_data,spaspm_data-method}: TODO
}}

