% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_class_definitions.R
\docType{class}
\name{spaspm_data-class}
\alias{spaspm_data-class}
\title{SPASPM dataset structure}
\description{
SPASPM dataset structure
}
\section{Slots}{

\describe{
\item{\code{data}}{\strong{[ANY]} TODO}

\item{\code{uniqueID}}{\strong{[character]} TODO}

\item{\code{is_spatial}}{\strong{[logical]} TODO}

\item{\code{coords}}{\strong{[character]} TODO}

\item{\code{representation}}{\strong{[character]} TODO}
}}

