% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_all_classes_definitions.R
\docType{class}
\name{spaspm_formula-class}
\alias{spaspm_formula-class}
\title{SPASPM formula object}
\description{
This class is a wrapper around the \code{formula} class. It is not intended for
users to directly manipulate and create new objects.
}
\section{Slots}{

\describe{
\item{\code{raw_formula}}{\strong{[formula]} The raw formula call}

\item{\code{translated_formula}}{\strong{[formula]} The translated formula call ready
to be evaluated.}

\item{\code{dataset}}{\strong{[character]} The name of the dataset the formula object is
to be mapped onto.}

\item{\code{vars}}{\strong{[list]} List of relevant variables for the evaluation of the
different smooths.}
}}

\seealso{
See the \code{mgcv} function for defining smooths: \link[mgcv:s]{s()}.
}
