% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm_smooth-accessors-methods.R
\name{spm_rep,spaspm_smooth-method}
\alias{spm_rep,spaspm_smooth-method}
\alias{spm_rep<-,spaspm_smooth-method}
\alias{spm_smooth,spaspm_smooth-method}
\alias{spm_smooth<-}
\alias{spm_smooth<-,spaspm_smooth-method}
\title{Accessing OR replacing \code{spaspm_smooth} object elements}
\usage{
\S4method{spm_rep}{spaspm_smooth}(spaspm_object)

\S4method{spm_rep}{spaspm_smooth}(object) <- value

\S4method{spm_smooth}{spaspm_smooth}(spaspm_object)

spm_smooth(object) <- value

\S4method{spm_smooth}{spaspm_smooth}(object) <- value
}
\arguments{
\item{spaspm_object}{\strong{[spaspm_smooth]} An object of class
\link[=spaspm_smooth-class]{spaspm_smooth}.}

\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=spaspm_smooth-class]{spaspm_smooth}.
}
\section{Functions}{
\itemize{
\item \code{spm_rep<-,spaspm_smooth-method}: TODO

\item \code{spm_smooth,spaspm_smooth-method}: TODO

\item \code{spm_smooth<-}: TODO

\item \code{spm_smooth<-,spaspm_smooth-method}: TODO
}}

