% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{spm}
\alias{spm}
\alias{spm,sspm,missing-method}
\alias{spm,sspm,formula-method}
\title{Fit an SPM model}
\usage{
spm(sspm_object, formula, ...)

\S4method{spm}{sspm,missing}(sspm_object, formula, ...)

\S4method{spm}{sspm,formula}(sspm_object, formula, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm_dataset]} An object of class
\link[=sspm_dataset-class]{sspm_dataset}.}

\item{formula}{\strong{[formula]} A formula definition of the form
response ~ smoothing_terms + ...}
}
\value{
The updated sspm object.
}
\description{
Fit an spm model to a sspm object
}
