% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_aggregate.R
\name{spm_aggregate}
\alias{spm_aggregate}
\alias{spm_aggregate,sspm_dataset,missing-method}
\alias{spm_aggregate,sspm_dataset,sspm_discrete_boundary-method}
\title{Aggregate a dataset variable based on a boundary}
\usage{
spm_aggregate(
  dataset,
  boundaries,
  level = "patch",
  type = "data",
  variable,
  fun,
  group_by = "spacetime",
  fill = FALSE,
  apply_to_df = FALSE,
  ...
)

\S4method{spm_aggregate}{sspm_dataset,missing}(
  dataset,
  boundaries,
  level = "patch",
  type = "data",
  variable,
  fun,
  group_by = "spacetime",
  fill = FALSE,
  apply_to_df = FALSE,
  ...
)

\S4method{spm_aggregate}{sspm_dataset,sspm_discrete_boundary}(
  dataset,
  boundaries,
  level = "patch",
  type = "data",
  variable,
  fun,
  group_by = "spacetime",
  fill = FALSE,
  apply_to_df = FALSE,
  ...
)
}
\arguments{
\item{dataset}{\strong{[sspm_dataset]} The dataset object.}

\item{boundaries}{\strong{[sspm_discrete_boundary]} The boundaries object
(optionnal).}

\item{variable}{\strong{[character]} Variable to aggregate (ignored in case
\code{apply_to_df} is \code{TRUE}).}

\item{fun}{\strong{[function]} Function to use to aggregate data.}

\item{group_by}{\strong{[character]} One of \code{time}, \code{space} and \code{spacetime}.}

\item{fill}{\strong{[logical OR numeric OR function]} Whether to complete the
incomplete cases, default to \code{FALSE} for no completion.}

\item{apply_to_df}{\strong{[logical]} Wether \code{fun} applied to  the data frame
group or to \code{variable}, default to \code{FALSE}.}

\item{...}{More arguments passed onto \code{fun}}
}
\value{
Updated \code{sspm_dataset}.
}
\description{
Aggregate the data contained in a dataset based on the discretized boundaries,
using a function and a filling value
}
