% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_as_boundary.R
\name{spm_as_boundary}
\alias{spm_as_boundary}
\alias{spm_as_boundary,sf,characterOrNULL-method}
\title{Create a sspm_boundary object}
\usage{
spm_as_boundary(boundaries, boundary_column = NULL)

\S4method{spm_as_boundary}{sf,characterOrNULL}(boundaries, boundary_column = NULL)
}
\arguments{
\item{boundaries}{\strong{[sf]} The sf object to cast.}

\item{boundary_column}{\strong{[character]} The column that comtains the possible
subdivisions of the boundaries.}
}
\description{
Create a sspm_boundary object.
}
