% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_as_dataset.R
\name{spm_as_dataset}
\alias{spm_as_dataset}
\alias{spm_as_dataset,data.frame,ANY,ANY,ANY,ANY,missingOrNULL-method}
\alias{spm_as_dataset,data.frame,ANY,ANY,ANY,ANY,list-method}
\alias{spm_as_dataset,data.frame,ANY,ANY,ANY,ANY,character-method}
\alias{spm_as_dataset,sf,ANY,ANY,ANY,ANY,ANY-method}
\title{Create a \code{sspm_dataset} dataset structure}
\usage{
spm_as_dataset(
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL
)

\S4method{spm_as_dataset}{data.frame,ANY,ANY,ANY,ANY,missingOrNULL}(
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL
)

\S4method{spm_as_dataset}{data.frame,ANY,ANY,ANY,ANY,list}(
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL
)

\S4method{spm_as_dataset}{data.frame,ANY,ANY,ANY,ANY,character}(
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL
)

\S4method{spm_as_dataset}{sf,ANY,ANY,ANY,ANY,ANY}(
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL
)
}
\arguments{
\item{data}{\strong{[data.frame OR sf]} The dataset.}

\item{name}{\strong{[character]} The name of the dataset, default to "Biomass".}

\item{type}{\strong{[character]} The type of dataset in the context of the sspm
workflow, one of "biomass", "predictor", or "catch".}

\item{time_column}{\strong{[character]} The column of \code{data} for the temporal
dimensions (i.e. year).}

\item{uniqueID}{\strong{[character]} The column of \code{data} that is unique for all
rows of the data matrix.}

\item{coords}{\strong{[character]} The column of \code{data} for longitude and
latitude of the observations.}

\item{crs}{Coordinate reference system, passed onto \link[=sf]{st_as_sf}.}
}
\value{
An object of class \code{\link[=sspm_dataset-class]{sspm_dataset}}.
}
\description{
This casts a \code{data.frame} or \code{sf} object into  an object of class
\code{\link[=sspm_dataset-class]{sspm_dataset}}.
}
