% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_discretize.R
\name{spm_discretize}
\alias{spm_discretize}
\alias{spm_discretize,ANY,ANY,missingOrNULL-method}
\alias{spm_discretize,ANY,missingOrNULL,ANY-method}
\alias{spm_discretize,ANY,ANY,character-method}
\alias{spm_discretize,sspm_boundary,sspm_dataset,discretization_method-method}
\alias{spm_discretize,sspm_boundary,sf,discretization_method-method}
\alias{spm_discretize,sspm_discrete_boundary,ANY,ANY-method}
\title{Discretize a \code{sspm} model object}
\usage{
spm_discretize(boundary_object, with, method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{ANY,ANY,missingOrNULL}(boundary_object, with, method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{ANY,missingOrNULL,ANY}(boundary_object, with, method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{ANY,ANY,character}(boundary_object, with, method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{sspm_boundary,sspm_dataset,discretization_method}(boundary_object, with, method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{sspm_boundary,sf,discretization_method}(boundary_object, with, method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{sspm_discrete_boundary,ANY,ANY}(boundary_object, with, method = "tesselate_voronoi", ...)
}
\arguments{
\item{boundary_object}{\strong{[sspm]} An object of class
\link[=sspm-class]{sspm_boundary}.}

\item{with}{\strong{[sspm_dataset OR sf]} Either an object of class sspm_dataset or
a set of custom points.}

\item{method}{\strong{[character OR method]}
Either a \code{character} from the list of available methods
(see \link{spm_methods} for the list) \strong{OR} an object of class
\link[=discretization_method-class]{discretization_method}.}

\item{...}{\strong{[named list]} Further arguments to be passed onto the function
used in \code{method}.}
}
\value{
An object of class \link[=sspm-class]{sspm_discrete_boundary} (the updated
and discretized \code{sspm} object given as input).
}
\description{
Discretize a \link[=sspm-class]{sspm} model object with a function from a
\link[=discretization_method-class]{discretization_method} object class.
}
