% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spm_methods}
\alias{spm_methods}
\title{Get the list of available discretization methods}
\usage{
spm_methods()
}
\value{
A \verb{character vector} of all available discretization methods.
}
\description{
Currently, only one discretization method is supported:
\itemize{
\item \code{"tesselate_voronoi"} Voronoi tessellation using the function
\link{tesselate_voronoi}.
}
}
\details{
You can create your own methid using TODO.
}
