% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_helpers.R
\name{spm_smooth_methods}
\alias{spm_smooth_methods}
\title{Get the list of available smoothing methods}
\usage{
spm_smooth_methods()
}
\value{
A \verb{character vector} of all available smoothing methods.
}
\description{
Currently, only one smoothing method is supported:
* \code{"ICAR"}: Intrinsic Conditional Auto-Regressive models.
* \code{"LINPRED"}: LINear PREDictors (lag smooths).
}
