% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_helpers.R
\name{spm_smooth_types}
\alias{spm_smooth_types}
\title{Get the list of available smooth types}
\usage{
spm_smooth_types()
}
\value{
A \verb{character vector} of all available smooth types.
}
\description{
Currently, only one smooth type is supported:
* \code{"ICAR"}
}
