% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspm.R
\name{sspm}
\alias{sspm}
\alias{sspm,sspm_dataset,sspm_dataset-method}
\alias{sspm,sspm_dataset,list-method}
\title{Create a \code{sspm} model object}
\usage{
sspm(biomass, predictors, catch = NULL, biomass_var = NULL, catch_var = NULL)

\S4method{sspm}{sspm_dataset,sspm_dataset}(biomass, predictors, catch = NULL, biomass_var = NULL, catch_var = NULL)

\S4method{sspm}{sspm_dataset,list}(biomass, predictors, catch = NULL, biomass_var = NULL, catch_var = NULL)
}
\arguments{
\item{name}{\strong{[character]} The name to be given to the model}
}
\value{
An object of class  \link[=sspm-class]{sspm}.
}
\description{
Create a sspm_model object.
}
