% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspm-package.R
\docType{package}
\name{sspm-package}
\alias{sspm-package}
\alias{_PACKAGE}
\title{sspm: Spatial Surplus Production Model Framework for Northern Shrimp Populations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The goal of `sspm` is to implement a gam-based spatial surplus 
    production model, aimed at modeling northern shrimp population in Canada but 
    potentially to any stock in any location. The package is opinionated in its 
    implementation of SPMs as it internally makes the choice to use penalized 
    spatial gams with time lags  based on Pedersen et al. (2020). However, it 
    also aims to provide options for the user to customize their model.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/pedersen-fisheries-lab/sspm/issues}
}

}
\author{
\strong{Maintainer}: Valentin Lucet \email{valentin.lucet@gmail.com}

Authors:
\itemize{
  \item Eric Pedersen \email{eric.pedersen@concordia.ca}
}

}
\keyword{internal}
