% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_S4_all_classes_definitions.R
\docType{class}
\name{sspm_boundary-class}
\alias{sspm_boundary-class}
\title{sspm boundary structure}
\description{
One of the first steps in the \code{sspm} workflow is to create one or more
object(s) of class \code{sspm_boundary} from an \code{sf} object.
}
\section{Slots}{

\describe{
\item{\code{boundaries}}{\strong{[sf]} Spatial boundaries (polygons).}

\item{\code{boundary}}{\strong{[character]} The column of \code{data} that represents the
spatial boundaries.}

\item{\code{boundary_area}}{\strong{[character]} The column of \code{data} that represents the
area of spatial boundaries.}
}}

