% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_all_classes_definitions.R
\docType{class}
\name{sspm_data-class}
\alias{sspm_data-class}
\title{sspm dataset structure}
\description{
One of the first step in the \code{sspm} workflow is to create one or more
object(s) of class \code{sspm_data} from a \code{data.frame}, \code{tibble} or \code{sf} object.
}
\section{Slots}{

\describe{
\item{\code{name}}{\strong{[character]} The name of the dataset, default to "Biomass".}

\item{\code{data}}{\strong{[data.frame OR sf]} The dataset.}

\item{\code{time_column}}{\strong{[character]} The column of \code{data} that represents the
temporal dimension of the dataset.}

\item{\code{coords}}{\strong{[character]} The columns of \code{data} that represent the
spatial dimension of the dataset: the two columns for longitude and
latitude of the observations.}

\item{\code{uniqueID}}{\strong{[character]} The column of \code{data} that is unique for all
rows of the data matrix.}

\item{\code{boundaries}}{\strong{[sf]} Spatial boundaries (polygons).}

\item{\code{formulas}}{\strong{[list]} \emph{(if discrete)} List of
\link[=sspm_formula-class]{sspm_formula} objects that are mapped onto the
base dataset.}

\item{\code{smoothed_data}}{\strong{[sf]} The smoothed data.}

\item{\code{smoothed_fit}}{\strong{[list]} The fit from smoothing the data}

\item{\code{is_split}}{\strong{[logical]} Whether or not this dataset has been splitted.}

\item{\code{is_mapped}}{\strong{[logical]} Whether the dataset has been mapped to
boundaries (used internally).}
}}

