% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_all_classes_definitions.R
\docType{class}
\name{sspm_fit-class}
\alias{sspm_fit-class}
\title{sspm fit}
\description{
The fit object for a sspm model
}
\section{Slots}{

\describe{
\item{\code{boundaries}}{\strong{[sf]} Spatial boundaries (polygons).}

\item{\code{datasets}}{\strong{[list]} \emph{(if discrete)} List of
\link[=sspm_data-class]{sspm_data} that define variables in the SPM model.}

\item{\code{smoothed_data}}{\strong{[sf]} The smoothed data.}
}}

