% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_S4_all_classes_definitions.R
\docType{class}
\name{sspm_formula-class}
\alias{sspm_formula-class}
\title{sspm formula object}
\description{
This class is a wrapper around the \code{formula} class. It is not intended for
users to directly manipulate and create new objects.
}
\section{Slots}{

\describe{
\item{\code{raw_formula}}{\strong{[formula]} The raw formula call}

\item{\code{translated_formula}}{\strong{[formula]} The translated formula call ready
to be evaluated.}

\item{\code{vars}}{\strong{[list]} List of relevant variables for the evaluation of the
different smooths.}

\item{\code{lag_vars}}{Smooth lag variables used for predictions}

\item{\code{response}}{\strong{[charatcer]} The response variable in the formula.}

\item{\code{is_fitted}}{\strong{[logical]} Whether this formula has already been fitted.}
}}

\seealso{
See the \code{mgcv} function for defining smooths: \link[mgcv:s]{s()}.
}
