% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-sspm_boundary.R
\name{spm_boundaries,sspm_boundary-method}
\alias{spm_boundaries,sspm_boundary-method}
\alias{spm_boundaries<-,sspm_boundary-method}
\alias{spm_discret_method}
\alias{spm_discret_method,sspm_discrete_boundary-method}
\alias{spm_discret_method<-}
\alias{spm_discret_method<-,sspm_discrete_boundary-method}
\alias{spm_patches}
\alias{spm_patches,sspm_discrete_boundary-method}
\alias{spm_patches<-}
\alias{spm_patches<-,sspm_discrete_boundary-method}
\alias{spm_points}
\alias{spm_points,sspm_discrete_boundary-method}
\alias{spm_points<-}
\alias{spm_points<-,sspm_discrete_boundary-method}
\alias{spm_boundary}
\alias{spm_boundary,sspm_boundary-method}
\alias{spm_boundary<-}
\alias{spm_boundary<-,sspm_boundary-method}
\alias{spm_boundary_area}
\alias{spm_boundary_area,sspm_boundary-method}
\alias{spm_boundary_area<-}
\alias{spm_boundary_area<-,sspm_boundary-method}
\alias{spm_patches_area}
\alias{spm_patches_area,sspm_discrete_boundary-method}
\alias{spm_patches_area<-}
\alias{spm_patches_area<-,sspm_discrete_boundary-method}
\title{Accessing OR replacing \code{sspm_boundary} model elements}
\usage{
\S4method{spm_boundaries}{sspm_boundary}(sspm_object)

\S4method{spm_boundaries}{sspm_boundary}(object) <- value

spm_discret_method(sspm_object)

\S4method{spm_discret_method}{sspm_discrete_boundary}(sspm_object)

spm_discret_method(object) <- value

\S4method{spm_discret_method}{sspm_discrete_boundary}(object) <- value

spm_patches(sspm_object)

\S4method{spm_patches}{sspm_discrete_boundary}(sspm_object)

spm_patches(object) <- value

\S4method{spm_patches}{sspm_discrete_boundary}(object) <- value

spm_points(sspm_object)

\S4method{spm_points}{sspm_discrete_boundary}(sspm_object)

spm_points(object) <- value

\S4method{spm_points}{sspm_discrete_boundary}(object) <- value

spm_boundary(sspm_object)

\S4method{spm_boundary}{sspm_boundary}(sspm_object)

spm_boundary(object) <- value

\S4method{spm_boundary}{sspm_boundary}(object) <- value

spm_boundary_area(sspm_object)

\S4method{spm_boundary_area}{sspm_boundary}(sspm_object)

spm_boundary_area(object) <- value

\S4method{spm_boundary_area}{sspm_boundary}(object) <- value

spm_patches_area(sspm_object)

\S4method{spm_patches_area}{sspm_discrete_boundary}(sspm_object)

spm_patches_area(object) <- value

\S4method{spm_patches_area}{sspm_discrete_boundary}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm_boundary]} An object of class
\link[=sspm_boundary-class]{sspm_boundary}.}

\item{object}{\strong{[sspm_boundary]} An object of class
\link[=sspm_boundary-class]{sspm_boundary}.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
The object in the required slot.
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm_boundary-class]{sspm_boundary}.
}
\examples{
data(borealis_simulated, package = "sspm")
biomass_dataset <- spm_as_dataset(data.frame(borealis_simulated), name = "borealis",
                                  density = "weight_per_km2",
                                  time = "year_f",
                                  coords = c('lon_dec','lat_dec'),
                                  uniqueID = "uniqueID")
spm_boundaries(biomass_dataset)
}
