% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-sspm_data.R
\name{spm_data}
\alias{spm_data}
\alias{spm_data,sspm-method}
\alias{spm_data,sspm_data-method}
\alias{spm_data<-}
\alias{spm_data<-,sspm_data-method}
\alias{spm_name,sspm_data-method}
\alias{spm_name<-,sspm_data-method}
\alias{spm_rep}
\alias{spm_rep,sspm_data-method}
\alias{spm_rep<-}
\alias{spm_rep<-,sspm_data-method}
\alias{spm_unique_ID,sspm_data-method}
\alias{spm_unique_ID<-,sspm_data-method}
\alias{spm_coords_col,sspm_data-method}
\alias{spm_coords_col<-,sspm_data-method}
\alias{spm_time_column,sspm_data-method}
\alias{spm_time_column<-,sspm_data-method}
\alias{spm_formulas,sspm_data-method}
\alias{spm_formulas<-,sspm_data-method}
\title{Accessing OR replacing \code{sspm_data} model elements}
\usage{
spm_data(sspm_object)

\S4method{spm_data}{sspm}(sspm_object)

\S4method{spm_data}{sspm_data}(sspm_object)

spm_data(object) <- value

\S4method{spm_data}{sspm_data}(object) <- value

\S4method{spm_name}{sspm_data}(sspm_object)

\S4method{spm_name}{sspm_data}(object) <- value

spm_rep(sspm_object)

\S4method{spm_rep}{sspm_data}(sspm_object)

spm_rep(object) <- value

\S4method{spm_rep}{sspm_data}(object) <- value

\S4method{spm_unique_ID}{sspm_data}(sspm_object)

\S4method{spm_unique_ID}{sspm_data}(object) <- value

\S4method{spm_coords_col}{sspm_data}(sspm_object)

\S4method{spm_coords_col}{sspm_data}(object) <- value

\S4method{spm_time_column}{sspm_data}(sspm_object)

\S4method{spm_time_column}{sspm_data}(object) <- value

\S4method{spm_formulas}{sspm_data}(sspm_object)

\S4method{spm_formulas}{sspm_data}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm_data]} An object of class
\link[=sspm_data-class]{sspm_data}.}

\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm_data-class]{sspm_data}.
}
\section{Functions}{
\itemize{
\item \code{spm_data,sspm-method}: TODO

\item \code{spm_data,sspm_data-method}: TODO

\item \code{spm_data<-}: TODO

\item \code{spm_data<-,sspm_data-method}: TODO

\item \code{spm_name,sspm_data-method}: TODO

\item \code{spm_name<-,sspm_data-method}: TODO

\item \code{spm_rep}: TODO

\item \code{spm_rep,sspm_data-method}: TODO

\item \code{spm_rep<-}: TODO

\item \code{spm_rep<-,sspm_data-method}: TODO

\item \code{spm_unique_ID,sspm_data-method}: TODO

\item \code{spm_unique_ID<-,sspm_data-method}: TODO

\item \code{spm_coords_col,sspm_data-method}: TODO

\item \code{spm_coords_col<-,sspm_data-method}: TODO

\item \code{spm_time_column,sspm_data-method}: TODO

\item \code{spm_time_column<-,sspm_data-method}: TODO

\item \code{spm_formulas,sspm_data-method}: TODO

\item \code{spm_formulas<-,sspm_data-method}: TODO
}}

