% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-methods-sspm_fit.R
\name{spm_unique_ID,sspm_fit-method}
\alias{spm_unique_ID,sspm_fit-method}
\alias{spm_unique_ID<-,sspm_fit-method}
\alias{spm_time,sspm_fit-method}
\alias{spm_time<-,sspm_fit-method}
\alias{spm_formulas,sspm_fit-method}
\alias{spm_formulas<-,sspm_fit-method}
\alias{spm_smoothed_data,sspm_fit-method}
\alias{spm_smoothed_data<-,sspm_fit-method}
\alias{spm_get_fit}
\alias{spm_get_fit,sspm_fit-method}
\alias{spm_get_fit<-}
\alias{spm_get_fit<-,sspm_fit-method}
\alias{spm_boundaries,sspm_fit-method}
\alias{spm_boundaries<-,sspm_fit-method}
\alias{spm_boundary,sspm_fit-method}
\alias{spm_boundary<-,sspm_fit-method}
\title{Accessing OR replacing \code{sspm_fit} model elements}
\usage{
\S4method{spm_unique_ID}{sspm_fit}(sspm_object)

\S4method{spm_unique_ID}{sspm_fit}(object) <- value

\S4method{spm_time}{sspm_fit}(sspm_object)

\S4method{spm_time}{sspm_fit}(object) <- value

\S4method{spm_formulas}{sspm_fit}(sspm_object)

\S4method{spm_formulas}{sspm_fit}(object) <- value

\S4method{spm_smoothed_data}{sspm_fit}(sspm_object)

\S4method{spm_smoothed_data}{sspm_fit}(object) <- value

spm_get_fit(sspm_object)

\S4method{spm_get_fit}{sspm_fit}(sspm_object)

spm_get_fit(object) <- value

\S4method{spm_get_fit}{sspm_fit}(object) <- value

\S4method{spm_boundaries}{sspm_fit}(sspm_object)

\S4method{spm_boundaries}{sspm_fit}(object) <- value

\S4method{spm_boundary}{sspm_fit}(sspm_object)

\S4method{spm_boundary}{sspm_fit}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm_fit]} An object of class
\link[=sspm_fit-class]{sspm_fit}.}

\item{object}{\strong{[sspm_fit]} An object of class
\link[=sspm_fit-class]{sspm_fit}.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
The object in the required slot.
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm_fit-class]{sspm_fit}.
}
\examples{
data(borealis_simulated, package = "sspm")
biomass_dataset <- spm_as_dataset(data.frame(borealis_simulated), name = "borealis",
                                  density = "weight_per_km2",
                                  time = "year_f",
                                  coords = c('lon_dec','lat_dec'),
                                  uniqueID = "uniqueID")
spm_formulas(biomass_dataset)
}
