% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_spaspm_smooth.R
\name{as_spaspm_smooth}
\alias{as_spaspm_smooth}
\alias{as_spaspm_smooth,spaspm_smooth,ANY,ANY-method}
\alias{as_spaspm_smooth,ANY,missing,ANY-method}
\alias{as_spaspm_smooth,character,character,character-method}
\alias{as_spaspm_smooth,ANY,character,character-method}
\title{Create a \code{spaspm_smooth} object}
\usage{
as_spaspm_smooth(smooth, dataset, dimension, ...)

\S4method{as_spaspm_smooth}{spaspm_smooth,ANY,ANY}(smooth)

\S4method{as_spaspm_smooth}{ANY,missing,ANY}(smooth, dataset, dimension, ...)

\S4method{as_spaspm_smooth}{character,character,character}(smooth, dataset, dimension, ...)

\S4method{as_spaspm_smooth}{ANY,character,character}(smooth, dataset, dimension, ...)
}
\arguments{
\item{smooth}{\strong{[mgcv xxx.smooth.spec object OR character]} A smooth
object from \link{mgcv} or a character value from the suppported default types. See
\link[=spm_smooth_types]{spm_smooth_types()}.}

\item{dataset}{\strong{[character]} The name of the dataset on which to
map the smooth. See also \link[=spaspm_smooth-class]{spaspm_smooth}.}

\item{dimension}{\strong{[character]} The smoothing dimension the smooth is
applied to: one if "space", "time", "space_time".}

\item{...}{Further arguments passed on to build the smooth \code{TODO}.}
}
\value{
An object of class \code{\link[=spaspm_smooth-class]{spaspm_smooth}}.
}
\description{
This function allows to cast different inputs into an object of class
\code{\link[=spaspm_smooth-class]{spaspm_smooth}}.
}
\section{Functions}{
\itemize{
\item \code{as_spaspm_smooth,spaspm_smooth,ANY,ANY-method}: TODO

\item \code{as_spaspm_smooth,ANY,missing,ANY-method}: TODO

\item \code{as_spaspm_smooth,character,character,character-method}: TODO

\item \code{as_spaspm_smooth,ANY,character,character-method}: TODO
}}

\seealso{
See the \code{mgcv} function for defining smooths: \link[mgcv:s]{s()}.
}
