% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_spaspm_smooth.R
\name{as_spaspm_smooth}
\alias{as_spaspm_smooth}
\alias{as_spaspm_smooth,spaspm_smooth,ANY,ANY-method}
\alias{as_spaspm_smooth,ANY,missing,ANY-method}
\alias{as_spaspm_smooth,ANY,character,missing-method}
\alias{as_spaspm_smooth,ANY,character,ANY-method}
\alias{as_spaspm_smooth,missing,character,character-method}
\title{Create a \code{spaspm_smooth} object}
\usage{
as_spaspm_smooth(smooth, dataset_name, type, ...)

\S4method{as_spaspm_smooth}{spaspm_smooth,ANY,ANY}(smooth)

\S4method{as_spaspm_smooth}{ANY,missing,ANY}(smooth, dataset_name, type, ...)

\S4method{as_spaspm_smooth}{ANY,character,missing}(smooth, dataset_name, type, ...)

\S4method{as_spaspm_smooth}{ANY,character,ANY}(smooth, dataset_name, type, ...)

\S4method{as_spaspm_smooth}{missing,character,character}(smooth, dataset_name, type, ...)
}
\arguments{
\item{smooth}{\strong{[mgcv xxx.smooth.spec object]} A smooth object from \link{mgcv}.}

\item{dataset_name}{\strong{[character]} The name of the dataset on which to
map the smooth. See also \link[=spaspm_smooth-class]{spaspm_smooth}.}

\item{type}{\strong{[character]}}
}
\value{
An object of class \code{\link[=spaspm_smooth-class]{spaspm_smooth}}.
}
\description{
This function allows to cast different inputs into an object of class
\code{\link[=spaspm_smooth-class]{spaspm_smooth}}.
}
\section{Functions}{
\itemize{
\item \code{as_spaspm_smooth,spaspm_smooth,ANY,ANY-method}: TODO

\item \code{as_spaspm_smooth,ANY,missing,ANY-method}: TODO

\item \code{as_spaspm_smooth,ANY,character,missing-method}: TODO

\item \code{as_spaspm_smooth,ANY,character,ANY-method}: TODO

\item \code{as_spaspm_smooth,missing,character,character-method}: TODO
}}

\seealso{
See the \code{mgcv} function for defining smooths: \link[mgcv:s]{s()}.
}
