% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dataset.R
\name{map_dataset}
\alias{map_dataset}
\alias{map_dataset,sspm,data.frame-method}
\alias{map_dataset,sspm,sf-method}
\alias{map_dataset,sspm,sspm_data-method}
\alias{map_dataset,sspm,list-method}
\alias{map_biomass,sspm,data.frame-method}
\alias{map_predictor,sspm,data.frame-method}
\alias{map_catch,sspm,data.frame-method}
\title{Map dataset onto a discretized sspm object}
\usage{
map_dataset(
  sspm_object,
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)

\S4method{map_dataset}{sspm,data.frame}(
  sspm_object,
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)

\S4method{map_dataset}{sspm,sf}(
  sspm_object,
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)

\S4method{map_dataset}{sspm,sspm_data}(
  sspm_object,
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)

\S4method{map_dataset}{sspm,list}(
  sspm_object,
  data,
  name,
  type,
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)

\S4method{map_biomass}{sspm,data.frame}(
  sspm_object,
  data,
  name,
  type = "biomass",
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)

\S4method{map_predictor}{sspm,data.frame}(
  sspm_object,
  data,
  name,
  type = "predictor",
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)

\S4method{map_catch}{sspm,data.frame}(
  sspm_object,
  data,
  name,
  type = "catch",
  time_column,
  uniqueID,
  coords = NULL,
  crs = NULL,
  ...
)
}
\arguments{
\item{sspm_object}{\strong{[sspm_discrete]} An object of class
\link[=sspm-class]{sspm} or \link[=sspm_discrete-class]{sspm_discrete}}

\item{data}{\strong{[data.frame OR sf]} The dataset.}

\item{name}{\strong{[character]} The name of the dataset, default to "Biomass".}

\item{type}{\strong{[character]} The type of dataset in the context of the sspm
workflow, one of "biomass", "predictor", or "catch".}

\item{time_column}{\strong{[character]} The column of \code{data} for the temporal
dimensions (i.e. year).}

\item{uniqueID}{\strong{[character]} The column of \code{data} that is unique for all
rows of the data matrix.}

\item{coords}{\strong{[character]} The column of \code{data} for longitude and
latitude of the observations.}

\item{crs}{Coordinate reference system, passed onto \link[=sf]{st_as_sf}.}

\item{...}{
  Arguments passed on to \code{\link[=as_sspm_data]{as_sspm_data}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
The updated object, of class \link[=sspm-class]{sspm} or
\link[=sspm_discrete-class]{sspm_discrete}.
}
\description{
After discretizing, the next step is to map coavriate datasets.
}
