% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dataset.R
\name{map_dataset}
\alias{map_dataset}
\alias{map_dataset,sspm_discrete,data.frame-method}
\alias{map_dataset,sspm_discrete,sf-method}
\alias{map_dataset,sspm_discrete,sspm_data-method}
\alias{map_dataset,sspm_discrete,list-method}
\alias{map_dataset,sspm,ANY-method}
\title{Map dataset onto a discretized sspm object}
\usage{
map_dataset(sspm_object, data, ...)

\S4method{map_dataset}{sspm_discrete,data.frame}(sspm_object, data, ...)

\S4method{map_dataset}{sspm_discrete,sf}(sspm_object, data, ...)

\S4method{map_dataset}{sspm_discrete,sspm_data}(sspm_object, data, ...)

\S4method{map_dataset}{sspm_discrete,list}(sspm_object, data, ...)

\S4method{map_dataset}{sspm,ANY}(sspm_object, data, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm_discrete]} An object of class
\link[=sspm-class]{sspm} or \link[=sspm_discrete-class]{sspm_discrete}}

\item{data}{\strong{[data.frame OR sf]} The dataset.}

\item{...}{
  Arguments passed on to \code{\link[=as_sspm_data]{as_sspm_data}}
  \describe{
    \item{\code{time_column}}{\strong{[character]} The column of \code{data} for the temporal
dimensions (i.e. year).}
    \item{\code{coords}}{\strong{[character]} The column of \code{data} for longitude and
latitude of the observations.}
    \item{\code{name}}{\strong{[character]} The name of the dataset, default to "Biomass".}
    \item{\code{uniqueID}}{\strong{[character]} The column of \code{data} that is unique for all
rows of the data matrix.}
    \item{\code{crs}}{Coordinate reference system, passed onto \link[=sf]{st_as_sf}.}
  }}
}
\value{
The updated object, of class \link[=sspm-class]{sspm} or
\link[=sspm_discrete-class]{sspm_discrete}.
}
\description{
After discretizing, the next step is to map coavriate datasets.
}
\section{Functions}{
\itemize{
\item \code{map_dataset,sspm_discrete,data.frame-method}: TODO

\item \code{map_dataset,sspm_discrete,sf-method}: TODO

\item \code{map_dataset,sspm_discrete,sspm_data-method}: TODO

\item \code{map_dataset,sspm_discrete,list-method}: TODO

\item \code{map_dataset,sspm,ANY-method}: TODO
}}

