% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_formula.R, R/spm_smooth.R
\name{map_formula}
\alias{map_formula}
\alias{map_formula,sspm_discrete,formula,missing-method}
\alias{map_formula,sspm_data,formula,ANY-method}
\alias{spm_smooth,sspm_data,missing-method}
\title{Map model formula onto a discretized sspm object}
\usage{
map_formula(sspm_object, formula, dataset, ...)

\S4method{map_formula}{sspm_discrete,formula,missing}(sspm_object, formula, dataset, ...)

\S4method{map_formula}{sspm_data,formula,ANY}(sspm_object, formula, dataset, ...)

\S4method{spm_smooth}{sspm_data,missing}(sspm_object, formula, keep_fit = FALSE, predict = TRUE, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm_data]} An object of class
\link[=sspm_discrete-class]{sspm_discrete}.}

\item{formula}{\strong{[formula]} A formula definition of the form
response ~ smoothing_terms.}
}
\value{
The updated object, of class \link[=sspm_discrete-class]{sspm_discrete}.
}
\description{
This functions is now used internally to map a formula onto a \code{sspm_data}
obejct.
}
