% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_formula.R
\name{map_formula}
\alias{map_formula}
\alias{map_formula,sspm,ANY,ANY-method}
\alias{map_formula,sspm_discrete,missing,ANY-method}
\alias{map_formula,sspm_discrete,formula,missing-method}
\alias{map_formula,sspm_discrete,formula,character-method}
\alias{map_formula,sspm_discrete,formula,sspm_data-method}
\title{Map model formula onto a discretized sspm object}
\usage{
map_formula(sspm_object, formula, dataset, ...)

\S4method{map_formula}{sspm,ANY,ANY}(sspm_object, formula, dataset, ...)

\S4method{map_formula}{sspm_discrete,missing,ANY}(sspm_object, formula, dataset, ...)

\S4method{map_formula}{sspm_discrete,formula,missing}(sspm_object, formula, dataset, ...)

\S4method{map_formula}{sspm_discrete,formula,character}(sspm_object, formula, dataset, ...)

\S4method{map_formula}{sspm_discrete,formula,sspm_data}(sspm_object, formula, dataset, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm_discrete]} An object of class
\link[=sspm_discrete-class]{sspm_discrete}.}

\item{formula}{\strong{[formula]} A formula definition of the form
response ~ smoothing_terms + ...}

\item{dataset}{\strong{[character]} The name(s) of the dataset(s) for which to
specify the formula.}

\item{...}{Further arguments passed down, none used for now.}
}
\value{
The updated object, of class \link[=sspm_discrete-class]{sspm_discrete}.
}
\description{
This functions allows to specify a model formula for a given discrete sspm
object. This formula makes use of specific smoothing terms \code{smooth_time()},
\code{smooth_space()}, \code{smooth_space_time()}. This formula can also contain fixed
effects and custom smooths, and can make use of specific smoothing terms
\code{smooth_time()}, \code{smooth_space()}, \code{smooth_space_time()}. See Details for
more explanations.
}
