% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_formula.R, R/spm_smooth.R
\name{map_formula}
\alias{map_formula}
\alias{map_formula,sspm_data,sspm_discrete_boundary,formula-method}
\alias{map_formula,sspm,missing,formula-method}
\alias{map_formula,sf,ANY,formula-method}
\alias{spm_smooth,ANY,missing,ANY-method}
\alias{spm_smooth,ANY,ANY,missing-method}
\alias{spm_smooth,ANY,ANY,sspm_boundary-method}
\title{Map model formula onto a sspm_data object}
\usage{
map_formula(sspm_object, boundaries, formula, time_column, ...)

\S4method{map_formula}{sspm_data,sspm_discrete_boundary,formula}(sspm_object, boundaries, formula, time_column, ...)

\S4method{map_formula}{sspm,missing,formula}(sspm_object, boundaries, formula, time_column, ...)

\S4method{map_formula}{sf,ANY,formula}(sspm_object, boundaries, formula, time_column, ...)

\S4method{spm_smooth}{ANY,missing,ANY}(
  sspm_object,
  formula,
  boundaries,
  keep_fit = FALSE,
  predict = TRUE,
  ...
)

\S4method{spm_smooth}{ANY,ANY,missing}(
  sspm_object,
  formula,
  boundaries,
  keep_fit = FALSE,
  predict = TRUE,
  ...
)

\S4method{spm_smooth}{ANY,ANY,sspm_boundary}(
  sspm_object,
  formula,
  boundaries,
  keep_fit = FALSE,
  predict = TRUE,
  ...
)
}
\arguments{
\item{sspm_object}{\strong{[sspm_data]} An object of class
\link[=sspm_data-class]{sspm_data}.}

\item{boundaries}{\strong{[sspm_boundary]} An object of class
\link[=sspm_boundary-class]{sspm_discrete_boundary}.}

\item{formula}{\strong{[formula]} A formula definition of the form
response ~ smoothing_terms + ...}

\item{keep_fit}{\strong{[logical]} Whether or not to keep the fitted values and
model (default to TRUE, set to FALSE to reduce memory footprint).}

\item{predict}{\strong{[logical]} Whether or not to generate the smoothed
predictions (necessary to fit the final SPM model, default to TRUE).}
}
\value{
The updated object.
}
\description{
This functions is now used internally to map a formula onto a \code{sspm_data}
or \code{sspm} object.
}
