% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_smooth.R
\name{map_smooth}
\alias{map_smooth}
\alias{map_smooth,spaspm-method}
\alias{map_smooth,spaspm_discrete-method}
\title{Specify a smoothing model for a discrete \code{spaspm} object}
\usage{
map_smooth(spaspm_object, time, space, space_time, ...)

\S4method{map_smooth}{spaspm}(spaspm_object, time, space, space_time, ...)

\S4method{map_smooth}{spaspm_discrete}(spaspm_object, time, space, space_time, ...)
}
\arguments{
\item{time}{TODO}

\item{space}{TODO}

\item{space_time}{TODO}
}
\value{
The updated object, of class \link[=spaspm_discrete-class]{spaspm_discrete}.
}
\description{
Allows to specify a smoothing model for a discrete \code{spaspm} object.
}
\section{Functions}{
\itemize{
\item \code{map_smooth,spaspm-method}: TODO

\item \code{map_smooth,spaspm_discrete-method}: TODO
}}

