% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspm-plot.R
\name{plot.sspm}
\alias{plot.sspm}
\alias{plot,sspm-method}
\alias{plot,sspm_discrete-method}
\title{Plot a discrete \code{sspm} model object}
\usage{
\S4method{plot}{sspm}(x)

\S4method{plot}{sspm_discrete}(x, smoothed_var = NULL, pages = "first", nrow = 3, ncol = 3)
}
\arguments{
\item{x}{\strong{[sspm_discrete]} An object of class
\link[=sspm_discrete-class]{sspm_discrete}.}
}
\value{
N/A
}
\description{
This will plot the patches and points of a \code{sspm_discrete} object.
}
