% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm-plot.R
\name{plot}
\alias{plot}
\alias{plot,spaspm_discrete,ANY-method}
\title{S4 plot method}
\usage{
\S4method{plot}{spaspm,ANY}(x)

\S4method{plot}{spaspm_discrete,ANY}(x)
}
\description{
S4 plot method
}
\section{Methods (by class)}{
\itemize{
\item \code{x = spaspm_discrete,y = ANY}: TODO
}}

