% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm-plot.R
\name{plot}
\alias{plot}
\alias{plot,spaspm_discrete-method}
\title{S4 plot method}
\usage{
\S4method{plot}{spaspm}(x)

\S4method{plot}{spaspm_discrete}(x)
}
\description{
S4 plot method
}
\section{Methods (by class)}{
\itemize{
\item \code{spaspm_discrete}: TODO
}}

