% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.sspm}
\alias{plot,sspm_boundary,missing-method}
\alias{plot,sspm_dataset,missing-method}
\alias{plot,sspm_fit,missing-method}
\title{Plot \code{sspm} objects}
\usage{
\S4method{plot}{sspm_boundary,missing}(x, y, ...)

\S4method{plot}{sspm_dataset,missing}(
  x,
  y,
  ...,
  smoothed_var = NULL,
  page = "first",
  nrow = 2,
  ncol = 4,
  log = TRUE
)

\S4method{plot}{sspm_fit,missing}(
  x,
  y,
  ...,
  smoothed_var = NULL,
  biomass = NULL,
  use_sf = TRUE,
  biomass_var = NULL,
  biomass_var_predict = NULL,
  biomass_var_smooth = NULL,
  biomass_var_origin = NULL,
  page = "first",
  nrow = 2,
  ncol = 4,
  log = TRUE
)
}
\arguments{
\item{x}{\strong{[sspm_...]} An object from this package.}

\item{y}{NOT USED (from generic).}

\item{...}{NOT USED (from generic).}

\item{smoothed_var}{\strong{[character]} Variable to plot.}

\item{page}{\strong{[character]} Either "first" for the firs page of plots, or
"all" for all pages}

\item{nrow}{\strong{[numeric]} The number of rows to paginate the plot on.}

\item{ncol}{\strong{[numeric]} The number of columns to paginate the plot on.}

\item{log}{\strong{[logical]} Whether to plot on a log scale, default to TRUE.}

\item{biomass}{\strong{[data.frame]} Biomass variable.}

\item{use_sf}{\strong{[logical]} Whether to produce a spatial plot.}

\item{biomass_var}{\strong{[character]} Biomass variable to plot.}

\item{biomass_var_predict}{\strong{[character]} Biomass variable to plot (from
predictions, optionnal).}

\item{biomass_var_smooth}{\strong{[character]} Biomass variable to plot (smooth,
optional).}

\item{biomass_var_origin}{\strong{[character]} Biomass variable to plot (from
original dataset, optionnal)}
}
\value{
A ggplot object.
}
\description{
Plot methods for a range of sspm objects.
}
