% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.sspm}
\alias{predict,sspm_fit-method}
\alias{predict,sspm_dataset-method}
\title{Predict with a SPM model}
\usage{
\S4method{predict}{sspm_fit}(
  object,
  new_data = NULL,
  biomass = NULL,
  aggregate = FALSE,
  interval = FALSE,
  next_ts = FALSE,
  type = "response"
)

\S4method{predict}{sspm_dataset}(object, new_data = NULL, discrete = TRUE, type = "response")
}
\arguments{
\item{object}{\emph{[sspm_fit]}* Fit object to predict from.}

\item{new_data}{\strong{[data.frame]} New data to predict with.}

\item{biomass}{\strong{[character]} Biomass variable.}

\item{aggregate}{\strong{[logical]} For biomass predictions only, whether to
aggregate the data to the boundary level. Default to FALSE.}

\item{interval}{\strong{[logical]} Whether or not to calculate confidence, and
when possible, prediction intervals.}

\item{next_ts}{\strong{[logical]} For biomass, predict next timestep.}
}
\value{
A \code{dataframe} of predictions.
}
\description{
Predict using a fitted SPM model on the whole data or on new data
}
