% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooths.R
\name{smooth_time}
\alias{smooth_time}
\alias{smooth_space}
\alias{smooth_space_time}
\alias{smooth_time,ANY,character,sspm_discrete-method}
\alias{smooth_space,ANY,character,sspm_discrete-method}
\alias{smooth_space_time,ANY,character,sspm_discrete-method}
\title{sspm Smoothing functions}
\usage{
smooth_time(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = "re",
  xt = NULL,
  ...
)

smooth_space(
  type = "ICAR",
  dataset,
  sspm_object,
  k = 30,
  bs = "mrf",
  xt = NULL,
  ...
)

smooth_space_time(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = c("re", "mrf"),
  xt = NULL,
  ...
)

\S4method{smooth_time}{ANY,character,sspm_discrete}(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = "re",
  xt = NULL,
  ...
)

\S4method{smooth_space}{ANY,character,sspm_discrete}(
  type = "ICAR",
  dataset,
  sspm_object,
  k = 30,
  bs = "mrf",
  xt = NULL,
  ...
)

\S4method{smooth_space_time}{ANY,character,sspm_discrete}(
  type = "ICAR",
  dataset,
  sspm_object,
  k = NULL,
  bs = c("re", "mrf"),
  xt = NULL,
  ...
)
}
\arguments{
\item{type}{\strong{[character]} Type of smooth, currently only "ICAR" is
supported.}

\item{dataset}{\strong{[character]} The name of the dataset among base and/or
mapped datasets for which to specify the formula}

\item{sspm_object}{\strong{[sspm_discrete]} An object of class
\link[=sspm_discrete-class]{sspm_discrete}.}

\item{k}{ the dimension of the basis used to represent the smooth term.
    The default depends on the number of variables that the smooth is a
    function of. \code{k} should not be less than the dimension of the
    null space of the penalty for the term (see
    \code{\link[mgcv]{null.space.dimension}}), but will be reset if
  it is. See \code{\link[mgcv]{choose.k}} for further information.}

\item{bs}{a two letter character string indicating the (penalized) smoothing basis to use.
(eg \code{"tp"} for thin plate regression spline, \code{"cr"} for cubic regression spline).
see \code{\link[mgcv]{smooth.terms}} for an over view of what is available.
}

\item{xt}{Any extra information required to set up a particular basis. Used
  e.g. to set large data set handling behaviour for \code{"tp"} basis.}

\item{...}{Further arguments passed down, none used for now.}
}
\value{
A list of 2 lists:
\itemize{
\item \code{args}, contains the arguments to be passed on to the mgcv smooths
\item \code{vars}, contains variables relevant to the evaluation of the smooth.
}
}
\description{
A full sspm formula contains calls to the smoothing terms \code{smooth_time()},
\code{smooth_space()}, \code{smooth_space_time()}.
}
