% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_all_classes_definitions.R
\docType{class}
\name{spaspm-class}
\alias{spaspm-class}
\alias{spaspm_discrete-class}
\title{SPASPM model classes}
\description{
The different model classes follow the typical workflow of \code{spaspm}:
\itemize{
\item \strong{\code{spaspm}} Basic model object.
\item \strong{\code{spaspm_discrete}} Discretized model object. Contains a
\link[=discretization_method-class]{discretization_method} object. It can also
containes "mapped datasets" (for example, predator or observator data).
}
}
\section{Functions}{
\itemize{
\item \code{spaspm_discrete-class}: spaspm_discrete
}}

\section{Slots}{

\describe{
\item{\code{name}}{\strong{[character]} Name of the model.}

\item{\code{data}}{\strong{[\link[=spaspm_data-class]{spaspm_data}]} Observationnal data.}

\item{\code{boundaries}}{\strong{[sf]} Spatial boundaries (polygons).}

\item{\code{method}}{\strong{[\link[=discretization_method-class]{discretization_method}]}
\emph{(if discrete)} discretization method used.}

\item{\code{patches}}{\strong{[sf]} \emph{(if discrete)} Patches resulting from
discretization.}

\item{\code{points}}{\strong{[sf]} \emph{(if discrete)} Sample points used for
discretization.}

\item{\code{mapped_datasets}}{\strong{[list]} \emph{(if discrete)} List of
\link[=spaspm_data-class]{spaspm_data} objects that are mapped ontp the
base dataset.}

\item{\code{mapped_formulas}}{\strong{[list]} \emph{(if discrete)} List of mapped formulas
used to specify a model.}
}}

