% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_class_definitions.R
\docType{class}
\name{spaspm-class}
\alias{spaspm-class}
\title{SPASPM model classes}
\description{
The different model classes follow the typical workflow of \code{spaspm}:
\itemize{
\item \strong{\code{spaspm}} Basic model object.
\item \strong{\code{spaspm_discrete}} Discretized model object. Contains a
\link[=discretization_method-class]{discretization_method} object.
}
}
\section{Slots}{

\describe{
\item{\code{name}}{\strong{[character]} Name of the model.}

\item{\code{data}}{\strong{[data.frame]} Observationnal data.}

\item{\code{boundaries}}{\strong{[sf]} Spatial boundaries (polygons).}

\item{\code{method}}{\strong{[\link[=discretization_method-class]{discretization_method}]}
\emph{(if discrete)} discretization method used.}

\item{\code{patches}}{\strong{[sf]} \emph{(if discrete)} Patches resulting from
discretization.}

\item{\code{points}}{\strong{[sf]} \emph{(if discrete)} Sample points used for
discretization.}
}}

