% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaspm_data-accessors-methods.R
\name{spm_data}
\alias{spm_data}
\alias{spm_data,spaspm-method}
\alias{spm_data,spaspm_data-method}
\alias{spm_data<-}
\alias{spm_data<-,spaspm_data-method}
\alias{spm_name,spaspm_data-method}
\alias{spm_name<-,spaspm_data-method}
\alias{spm_rep,spaspm_data-method}
\alias{spm_rep<-}
\alias{spm_rep<-,spaspm_data-method}
\title{Accessing OR replacing \code{spaspm_data} model elements}
\usage{
spm_data(spaspm_object)

\S4method{spm_data}{spaspm}(spaspm_object)

\S4method{spm_data}{spaspm_data}(spaspm_object)

spm_data(object) <- value

\S4method{spm_data}{spaspm_data}(object) <- value

\S4method{spm_name}{spaspm_data}(spaspm_object)

\S4method{spm_name}{spaspm_data}(object) <- value

\S4method{spm_rep}{spaspm_data}(spaspm_object)

spm_rep(object) <- value

\S4method{spm_rep}{spaspm_data}(object) <- value
}
\arguments{
\item{spaspm_object}{\strong{[spaspm_data]} An object of class
\link[=spaspm_data-class]{spaspm_data}.}

\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=spaspm_data-class]{spaspm_data}.
}
\section{Functions}{
\itemize{
\item \code{spm_data,spaspm-method}: TODO

\item \code{spm_data,spaspm_data-method}: TODO

\item \code{spm_data<-}: TODO

\item \code{spm_data<-,spaspm_data-method}: TODO

\item \code{spm_name,spaspm_data-method}: TODO

\item \code{spm_name<-,spaspm_data-method}: TODO

\item \code{spm_rep,spaspm_data-method}: TODO

\item \code{spm_rep<-}: TODO

\item \code{spm_rep<-,spaspm_data-method}: TODO
}}

