% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_class_definitions.R
\docType{class}
\name{spaspm_data-class}
\alias{spaspm_data-class}
\title{SPASPM dataset structure}
\description{
The first step in the \code{spaspm} workflow is to register the base dataset
(usually biomass) using the \link{spaspm} function.
}
\section{Slots}{

\describe{
\item{\code{name}}{\strong{[character]} The name of the dataset, default to "Biomass".}

\item{\code{data}}{\strong{[data.frame OR sf]} The dataset.}

\item{\code{uniqueID}}{\strong{[character]} The column of \code{data} that is unique for all
rows of the data matrix.}

\item{\code{coords}}{\strong{[character]} The column of \code{data} for longitude and
latitude of the observations.}

\item{\code{representation}}{\strong{[character]} Used internally and for print methods,
encodes the type of dataset.}
}}

