% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_all_classes_definitions.R
\docType{class}
\name{spaspm_smooth-class}
\alias{spaspm_smooth-class}
\title{SPASPM smooth object}
\description{
This class is a wrapper around the \code{xxx.smooth.spec} classes in \link{mgcv}.
These S3 classes represents all possibe smooths that can be fitted in the
package. It is used internally to formally cast users inputs for smooths
specifications. It is not directly intended for the user to instantiate
objects, but it is still exported.
}
\section{Slots}{

\describe{
\item{\code{representation}}{\strong{[character]} A name for the way the user specified
the smooths input.}

\item{\code{dataset_name}}{\strong{[character]} The name of the dataset the smooth object is
to be mapped onto. Will be checked against the list of datasets in the
SPASPM object when \link{map_smooth} is called.}

\item{\code{dimension}}{\strong{[character]} The smoothing dimension the smooth is
applied to: one if "space", "time", "space_time".}

\item{\code{smooth}}{\strong{[xxx.smooth.spec]} An object of class \code{xxx.smooth.spec}.}
}}

\seealso{
See the \code{mgcv} function for defining smooths: \link[mgcv:s]{s()}.
}
