% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_aggregate.R
\name{spm_aggregate}
\alias{spm_aggregate}
\alias{spm_aggregate,sspm_dataset,sspm_discrete_boundaryOrNULL-method}
\alias{spm_aggregate_catch,sspm_dataset,sspm_dataset,character,character-method}
\title{Aggregate a dataset variable based on a boundary}
\usage{
spm_aggregate(
  dataset,
  boundaries = NULL,
  variable,
  fun,
  group_by = "spacetime",
  fill = FALSE,
  apply_to_df = FALSE,
  ...
)

\S4method{spm_aggregate}{sspm_dataset,sspm_discrete_boundaryOrNULL}(
  dataset,
  boundaries = NULL,
  variable,
  fun,
  group_by = "spacetime",
  fill = FALSE,
  apply_to_df = FALSE,
  ...
)

\S4method{spm_aggregate_catch}{sspm_dataset,sspm_dataset,character,character}(
  biomass,
  catch,
  biomass_variable,
  catch_variable,
  fun = sum,
  group_by = "spacetime",
  fill,
  apply_to_df = FALSE,
  ...
)
}
\arguments{
\item{dataset}{\strong{[sspm_dataset]} The dataset object.}

\item{boundaries}{\strong{[sspm_discrete_boundary]} The boundaries object
(optionnal).}

\item{variable}{\strong{[character]} Variable to aggregate (ignored in case
\code{apply_to_df} is \code{TRUE}).}

\item{fun}{\strong{[function]} Function to use to aggregate data.}

\item{group_by}{\strong{[character]} One of \code{time}, \code{space} and \code{spacetime}.}

\item{fill}{\strong{[logical OR numeric OR function]} Whether to complete the
incomplete cases, default to \code{FALSE} for no completion.}

\item{apply_to_df}{\strong{[logical]} Wether \code{fun} applied to  the data frame
group or to \code{variable}, default to \code{FALSE}.}

\item{...}{More arguments passed onto \code{fun}}

\item{biomass}{\strong{[sspm_dataset (smoothed)]} The dataset containing the
biomass variable.}

\item{catch}{\strong{[sspm_dataset]} The dataset containing the catch variable.}

\item{biomass_variable}{\strong{[character]} The biomass variab of \code{biomass}.}

\item{catch_variable}{\strong{[character]} The catch column of \code{catch}.}
}
\description{
Aggregate the data contained in a dataset based on the discretized boundaries,
using a function and a filling value
}
\section{Functions}{
\itemize{
\item \code{spm_aggregate_catch,sspm_dataset,sspm_dataset,character,character-method}: spm_aggregate_catch
}}

