% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_as_boundary.R
\name{spm_as_boundary}
\alias{spm_as_boundary}
\alias{spm_as_boundary,sf,NULL,ANY-method}
\alias{spm_as_boundary,sf,character,missing-method}
\alias{spm_as_boundary,sf,character,character-method}
\title{Create a sspm_boundary object}
\usage{
spm_as_boundary(boundaries, boundary_column, surface_column)

\S4method{spm_as_boundary}{sf,`NULL`,ANY}(boundaries, boundary_column, surface_column)

\S4method{spm_as_boundary}{sf,character,missing}(boundaries, boundary_column, surface_column)

\S4method{spm_as_boundary}{sf,character,character}(boundaries, boundary_column, surface_column)
}
\arguments{
\item{boundaries}{\strong{[sf]} The sf object to cast.}

\item{boundary_column}{\strong{[character]} The column that contains the possible
subdivisions of the boundaries.}

\item{surface_column}{\strong{[character]} The column that contains the surface
of the subdivisions (optional).}
}
\description{
Create a sspm_boundary object.
}
