% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_as_boundary.R
\name{spm_as_boundary}
\alias{spm_as_boundary}
\alias{spm_as_boundary,sf,NULL,ANY,ANY,ANY-method}
\alias{spm_as_boundary,sf,character,missing,missing,missing-method}
\alias{spm_as_boundary,sf,character,character,missing,missing-method}
\alias{spm_as_boundary,sf,character,missing,ANY,ANY-method}
\alias{spm_as_boundary,sf,character,character,ANY,ANY-method}
\title{Create a sspm_boundary object}
\usage{
spm_as_boundary(
  boundaries,
  boundary_column,
  surface_column,
  patches = NULL,
  points = NULL,
  ...
)

\S4method{spm_as_boundary}{sf,`NULL`,ANY,ANY,ANY}(boundaries, boundary_column, surface_column)

\S4method{spm_as_boundary}{sf,character,missing,missing,missing}(boundaries, boundary_column, surface_column)

\S4method{spm_as_boundary}{sf,character,character,missing,missing}(boundaries, boundary_column, surface_column)

\S4method{spm_as_boundary}{sf,character,missing,ANY,ANY}(boundaries, boundary_column, surface_column, patches, points)

\S4method{spm_as_boundary}{sf,character,character,ANY,ANY}(boundaries, boundary_column, surface_column, patches, points)
}
\arguments{
\item{boundaries}{\strong{[sf]} The sf object to cast.}

\item{boundary_column}{\strong{[character]} The column that contains the possible
subdivisions of the boundaries.}

\item{surface_column}{\strong{[character]} The column that contains the surface
of the subdivisions (optional).}

\item{patches}{\strong{[sf]} Patches resulting from discretization.}

\item{points}{\strong{[sf]} Sample points used for discretization.}

\item{...}{further args passed onto methods}
}
\value{
An object of class \link[=sspm_boundary-class]{sspm_boundary} or
\link[=sspm_discrete_boundary-class]{sspm_discrete_boundary}.
}
\description{
Create a sspm_boundary object.
}
