% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_discretize.R
\name{spm_discretize}
\alias{spm_discretize}
\alias{spm_discretize,sspm,character,discretization_method-method}
\alias{spm_discretize,sspm,character,character-method}
\alias{spm_discretize,sspm,ANY,NULL-method}
\alias{spm_discretize,sspm_discrete,character,character-method}
\alias{spm_discretize,sspm_discrete,character,discretization_method-method}
\title{Discretize a \code{sspm} model object}
\usage{
spm_discretize(
  sspm_object,
  with_dataset,
  discretization_method = "tesselate_voronoi",
  ...
)

\S4method{spm_discretize}{sspm,character,discretization_method}(
  sspm_object,
  with_dataset,
  discretization_method = "tesselate_voronoi",
  ...
)

\S4method{spm_discretize}{sspm,character,character}(
  sspm_object,
  with_dataset,
  discretization_method = "tesselate_voronoi",
  ...
)

\S4method{spm_discretize}{sspm,ANY,`NULL`}(
  sspm_object,
  with_dataset,
  discretization_method = "tesselate_voronoi",
  ...
)

\S4method{spm_discretize}{sspm_discrete,character,character}(
  sspm_object,
  with_dataset,
  discretization_method = "tesselate_voronoi",
  ...
)

\S4method{spm_discretize}{sspm_discrete,character,discretization_method}(
  sspm_object,
  with_dataset,
  discretization_method,
  force = FALSE,
  ...
)
}
\arguments{
\item{sspm_object}{\strong{[sspm]} An object of class
\link[=sspm-class]{sspm}.}

\item{with_dataset}{\strong{[character]} The name of the dataset to use for the
discretization.}

\item{discretization_method}{\strong{[character OR discretization_method]}
Either a \code{character} from the list of available methods
(see \link{spm_methods} for the list) \strong{OR} an object of class
\link[=discretization_method-class]{discretization_method}.}

\item{...}{\strong{[named list]} Further arguments to be passed onto the function
used in the \code{discretization_method}.}

\item{force}{\strong{[boolean]} Only used when calling \code{spm_discretize} onto
an object of class \link[=sspm-class]{sspm_discrete}: whether you wish
to force the re-discretization of that model.}
}
\value{
An object of class \link[=sspm-class]{sspm_discrete} (the updated
and discretized \code{sspm} object given as input).
}
\description{
Discretize a \link[=sspm-class]{sspm} model object with a function from a
\link[=discretization_method-class]{discretization_method} object class.
}
\section{Functions}{
\itemize{
\item \code{spm_discretize,sspm,character,discretization_method-method}: TODO

\item \code{spm_discretize,sspm,character,character-method}: TODO

\item \code{spm_discretize,sspm,ANY,NULL-method}: TODO

\item \code{spm_discretize,sspm_discrete,character,character-method}: TODO

\item \code{spm_discretize,sspm_discrete,character,discretization_method-method}: TODO
}}

