% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_discretize.R
\name{spm_discretize}
\alias{spm_discretize}
\alias{spm_discretize,spaspm,discretization_method-method}
\alias{spm_discretize,spaspm,character-method}
\alias{spm_discretize,spaspm_discrete,character-method}
\alias{spm_discretize,spaspm_discrete,discretization_method-method}
\title{Discretize a \code{spaspm} model object}
\usage{
spm_discretize(spaspm_object, discretization_method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{spaspm,discretization_method}(spaspm_object, discretization_method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{spaspm,character}(spaspm_object, discretization_method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{spaspm_discrete,character}(spaspm_object, discretization_method = "tesselate_voronoi", ...)

\S4method{spm_discretize}{spaspm_discrete,discretization_method}(spaspm_object, discretization_method, force = FALSE, ...)
}
\arguments{
\item{spaspm_object}{\strong{[spaspm]} An object of class
\link[=spaspm-class]{spaspm}.}

\item{discretization_method}{\strong{[character OR discretization_method]}
Either a \code{character} from the list of available methods
(see \link{spm_methods} for the list) \strong{OR} an object of class
\link[=discretization_method-class]{discretization_method}.}

\item{...}{\strong{[named list]} Further arguments to be passed onto the function
used in the \code{discretization_method}.}

\item{force}{\strong{[boolean]} Only used when calling \code{spm_discretize} onto
an object of class \link[=spaspm-class]{spaspm_discrete}: whether you wish
to force the re-discretization of that model.}
}
\value{
An object of class \link[=spaspm-class]{spaspm_discrete} (the updated
and discretized \code{spaspm} object given as input).
}
\description{
Discretize a \link[=spaspm-class]{spaspm} model object with a function from a
\link[=discretization_method-class]{discretization_method} object class.
}
\section{Functions}{
\itemize{
\item \code{spm_discretize,spaspm,discretization_method-method}: TODO

\item \code{spm_discretize,spaspm,character-method}: TODO

\item \code{spm_discretize,spaspm_discrete,character-method}: TODO

\item \code{spm_discretize,spaspm_discrete,discretization_method-method}: TODO
}}

