% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_gam.R
\name{spm_gam}
\alias{spm_gam}
\alias{spm_gam,sspm_discrete-method}
\alias{spm_gam,sspm-method}
\alias{spm_gam,sspm_gam_fit-method}
\title{Fit spatial gams}
\usage{
spm_gam(sspm_object, ...)

\S4method{spm_gam}{sspm_discrete}(sspm_object, ...)

\S4method{spm_gam}{sspm}(sspm_object, ...)

\S4method{spm_gam}{sspm_gam_fit}(sspm_object, force = FALSE, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm_discrete]} An object of class
\link[=sspm_discrete-class]{sspm_discrete}.}

\item{...}{TBD}

\item{force}{\strong{[boolean]} Only used when calling \code{spm_discretize} onto
an object of class \link[=sspm-class]{sspm_discrete}: whether you wish
to force the re-discretization of that model.}
}
\value{
TBD.
}
\description{
TBD.
}
\section{Functions}{
\itemize{
\item \code{spm_gam,sspm_discrete-method}: TODO

\item \code{spm_gam,sspm-method}: TODO

\item \code{spm_gam,sspm_gam_fit-method}: TODO
}}

