% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_gam.R
\name{spm_gam}
\alias{spm_gam}
\alias{spm_gam,spaspm_discrete-method}
\alias{spm_gam,spaspm-method}
\alias{spm_gam,spaspm_gam_fit-method}
\title{Fit spatial gams}
\usage{
spm_gam(spaspm_object, ...)

\S4method{spm_gam}{spaspm_discrete}(spaspm_object, ...)

\S4method{spm_gam}{spaspm}(spaspm_object, ...)

\S4method{spm_gam}{spaspm_gam_fit}(spaspm_object, force = FALSE, ...)
}
\description{
Fit spatial gams
}
\section{Functions}{
\itemize{
\item \code{spm_gam,spaspm_discrete-method}: TODO

\item \code{spm_gam,spaspm-method}: TODO

\item \code{spm_gam,spaspm_gam_fit-method}: TODO
}}

