% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_plot.R
\name{spm_plot}
\alias{spm_plot}
\title{Plot various \code{sspm} objects}
\usage{
spm_plot(
  sspm_object,
  smoothed_var = NULL,
  page = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{sspm_object}{An object from this package.}

\item{smoothed_var}{TODO}

\item{page}{TODO}

\item{nrow}{TODO}

\item{ncol}{TODO}
}
\value{
N/A
}
\description{
Plot methods for a range of sspm objects.
}
