% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{spm_predict}
\alias{spm_predict}
\alias{spm_predict,sspm_discrete-method}
\title{Predict with a SPM model}
\usage{
spm_predict(sspm_object, new_data = NULL, ...)

\S4method{spm_predict}{sspm_discrete}(sspm_object, new_data = NULL, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm_discrete]} An object of class
\link[=sspm-class]{sspm} or \link[=sspm_discrete-class]{sspm_discrete}.}

\item{new_data}{\strong{[data.frame]}}

\item{...}{Arguments passed on to \link{predict.bam}.}
}
\description{
Predict using a fitted SPM model on the whole data or on new data
}
