% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{spm_predict}
\alias{spm_predict}
\alias{spm_predict,sspm_fit,data.frame-method}
\alias{spm_predict,sspm_fit,missing-method}
\alias{spm_predict,sspm_fit,list-method}
\title{Predict with a SPM model}
\usage{
spm_predict(sspm_object, new_data = NULL, ...)

\S4method{spm_predict}{sspm_fit,data.frame}(sspm_object, new_data = NULL, ...)

\S4method{spm_predict}{sspm_fit,missing}(sspm_object, new_data = NULL, ...)

\S4method{spm_predict}{sspm_fit,list}(sspm_object, new_data = NULL, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm_data]} An object of class
\link[=sspm_data-class]{sspm_data}.}

\item{new_data}{\strong{[data.frame]}}

\item{...}{Arguments passed on to \link{predict.bam}.}
}
\description{
Predict using a fitted SPM model on the whole data or on new data
}
