% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_split.R
\name{spm_split}
\alias{spm_split}
\alias{spm_split,sspm-method}
\title{Split data in test and train sets}
\usage{
spm_split(sspm_object, ...)

\S4method{spm_split}{sspm}(sspm_object, ...)
}
\arguments{
\item{sspm_object}{\strong{[sspm]} An object of class \link[=sspm-class]{sspm}.}

\item{...}{\strong{[expression]} Expression to evaluate to split data.}
}
\value{
The updated sspm object.
}
\description{
Split data before fitting spm.
}
\examples{
\dontrun{
sspm_model <- sspm_model \%>\%
    spm_split(year_f \%in\% c(1990:2017))
}

}
