% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_split.R
\name{spm_split}
\alias{spm_split}
\alias{spm_split,sspm,ANY-method}
\alias{spm_split,sspm_discrete,missing-method}
\alias{spm_split,sspm_discrete,character-method}
\title{Split data in test and train sets}
\usage{
spm_split(sspm_object, dataset, predicate)

\S4method{spm_split}{sspm,ANY}(sspm_object)

\S4method{spm_split}{sspm_discrete,missing}(sspm_object, dataset, predicate)

\S4method{spm_split}{sspm_discrete,character}(sspm_object, dataset, predicate)
}
\description{
Split data before fitting spm.
}
