% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_split.R
\name{spm_split}
\alias{spm_split}
\alias{spm_split,sspm,ANY-method}
\alias{spm_split,sspm_discrete,missing-method}
\alias{spm_split,sspm_discrete,character-method}
\title{Split data in test and train sets}
\usage{
spm_split(sspm_object, dataset, predicate)

\S4method{spm_split}{sspm,ANY}(sspm_object)

\S4method{spm_split}{sspm_discrete,missing}(sspm_object, dataset, predicate)

\S4method{spm_split}{sspm_discrete,character}(sspm_object, dataset, predicate)
}
\arguments{
\item{sspm_object}{\strong{[sspm_discrete]} An object of class
\link[=sspm-class]{sspm} or \link[=sspm_discrete-class]{sspm_discrete}}

\item{dataset}{\strong{[expression]} The dataset to split.}

\item{predicate}{\strong{[expression]} Expression to evaluate to split data.}
}
\value{
The updated sspm object.
}
\description{
Split data before fitting spm (WIP).
}
