% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_accessors-methods.R
\name{spm_name}
\alias{spm_name}
\alias{spm_}
\alias{spm_name<-}
\alias{spm_datasets}
\alias{spm_datasets,sspm-method}
\alias{spm_datasets<-}
\alias{spm_datasets<-,sspm-method}
\alias{spm_boundaries}
\alias{spm_boundaries,sspm-method}
\alias{spm_boundaries<-}
\alias{spm_boundaries<-,sspm-method}
\alias{spm_smoothed_data}
\alias{spm_smoothed_data,sspm-method}
\alias{spm_smoothed_data<-}
\alias{spm_smoothed_data<-,sspm-method}
\alias{spm_time}
\alias{spm_time,sspm-method}
\alias{spm_time<-}
\alias{spm_time<-,sspm-method}
\alias{is_split}
\alias{is_split,sspm-method}
\alias{is_split<-}
\alias{is_split<-,sspm-method}
\alias{spm_unique_ID}
\alias{spm_unique_ID,sspm-method}
\alias{spm_unique_ID<-}
\alias{spm_unique_ID<-,sspm-method}
\title{Accessing OR replacing \code{sspm} model elements}
\usage{
spm_name(sspm_object)

spm_name(object) <- value

spm_datasets(sspm_object)

\S4method{spm_datasets}{sspm}(sspm_object)

spm_datasets(object) <- value

\S4method{spm_datasets}{sspm}(object) <- value

spm_boundaries(sspm_object)

\S4method{spm_boundaries}{sspm}(sspm_object)

spm_boundaries(object) <- value

\S4method{spm_boundaries}{sspm}(object) <- value

spm_smoothed_data(sspm_object)

\S4method{spm_smoothed_data}{sspm}(sspm_object)

spm_smoothed_data(object) <- value

\S4method{spm_smoothed_data}{sspm}(object) <- value

spm_time(sspm_object)

\S4method{spm_time}{sspm}(sspm_object)

spm_time(object) <- value

\S4method{spm_time}{sspm}(object) <- value

is_split(sspm_object)

\S4method{is_split}{sspm}(sspm_object)

is_split(object) <- value

\S4method{is_split}{sspm}(object) <- value

spm_unique_ID(sspm_object)

\S4method{spm_unique_ID}{sspm}(sspm_object)

spm_unique_ID(object) <- value

\S4method{spm_unique_ID}{sspm}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm OR adjacent]} An object of class
\link[=sspm-class]{sspm} or others derivative classes.}

\item{object}{\strong{[sspm OR adjacent]} An object of class
\link[=sspm-class]{sspm} or others derivative classes.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\value{
The object in the required slot.
}
\description{
All methods described here allow to access the elements of contained in
objects of the different classes of the package.
}
\examples{
data(borealis_simulated, package = "sspm")
biomass_dataset <- spm_as_dataset(data.frame(borealis_simulated), name = "borealis",
                                  density = "weight_per_km2",
                                  time = "year_f",
                                  coords = c('lon_dec','lat_dec'),
                                  uniqueID = "uniqueID")
spm_name(biomass_dataset)
}
