% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_accessors-methods.R
\name{spm_name}
\alias{spm_name}
\alias{spm_}
\alias{spm_name,sspm-method}
\alias{spm_name<-}
\alias{spm_name<-,sspm-method}
\alias{spm_base_dataset}
\alias{spm_base_dataset,sspm-method}
\alias{spm_base_dataset,sspm_discrete-method}
\alias{spm_base_dataset<-}
\alias{spm_base_dataset<-,sspm-method}
\alias{spm_datasets}
\alias{spm_datasets,sspm-method}
\alias{spm_datasets,sspm_discrete-method}
\alias{spm_datasets,sspm_data-method}
\alias{spm_unique_ID}
\alias{spm_unique_ID,sspm-method}
\alias{spm_unique_ID,sspm_discrete-method}
\alias{spm_unique_ID<-}
\alias{spm_unique_ID<-,sspm-method}
\alias{spm_coords_col}
\alias{spm_coords_col,sspm-method}
\alias{spm_coords_col,sspm_discrete-method}
\alias{spm_coords_col<-}
\alias{spm_coords_col<-,sspm-method}
\alias{spm_time_col}
\alias{spm_time_col,sspm-method}
\alias{spm_time_col,sspm_discrete-method}
\alias{spm_time_col<-}
\alias{spm_time_col<-,sspm-method}
\alias{spm_boundaries}
\alias{spm_boundaries,sspm-method}
\alias{spm_boundaries<-}
\alias{spm_boundaries<-,sspm_discrete-method}
\alias{spm_boundaries<-,sspm-method}
\alias{spm_discret_method}
\alias{spm_discret_method,sspm-method}
\alias{spm_discret_method,sspm_discrete-method}
\alias{spm_discret_method<-}
\alias{spm_discret_method<-,sspm_discrete-method}
\alias{spm_patches}
\alias{spm_patches,sspm-method}
\alias{spm_patches,sspm_discrete-method}
\alias{spm_patches<-}
\alias{spm_patches<-,sspm_discrete-method}
\alias{spm_patches<-,sspm-method}
\alias{spm_points}
\alias{spm_points,sspm-method}
\alias{spm_points,sspm_discrete-method}
\alias{spm_points<-}
\alias{spm_points<-,sspm_discrete-method}
\alias{spm_points<-,sspm-method}
\alias{spm_mapped_datasets}
\alias{spm_mapped_datasets,sspm-method}
\alias{spm_mapped_datasets,sspm_discrete-method}
\alias{spm_mapped_datasets<-}
\alias{spm_mapped_datasets<-,sspm_discrete-method}
\alias{spm_mapped_datasets<-,sspm-method}
\alias{spm_mapped_formulas}
\alias{spm_mapped_formulas,sspm-method}
\alias{spm_mapped_formulas,sspm_discrete-method}
\alias{spm_mapped_formulas<-}
\alias{spm_mapped_formulas<-,sspm_discrete-method}
\alias{spm_mapped_formulas<-,sspm-method}
\title{Accessing OR replacing \code{sspm} model elements}
\usage{
spm_name(sspm_object)

\S4method{spm_name}{sspm}(sspm_object)

spm_name(object) <- value

\S4method{spm_name}{sspm}(object) <- value

spm_base_dataset(sspm_object)

\S4method{spm_base_dataset}{sspm}(sspm_object)

\S4method{spm_base_dataset}{sspm_discrete}(sspm_object)

spm_base_dataset(object) <- value

\S4method{spm_base_dataset}{sspm}(object) <- value

spm_datasets(sspm_object)

\S4method{spm_datasets}{sspm}(sspm_object)

\S4method{spm_datasets}{sspm_discrete}(sspm_object)

\S4method{spm_datasets}{sspm_data}(sspm_object)

spm_unique_ID(sspm_object)

\S4method{spm_unique_ID}{sspm}(sspm_object)

\S4method{spm_unique_ID}{sspm_discrete}(sspm_object)

spm_unique_ID(object) <- value

\S4method{spm_unique_ID}{sspm}(object) <- value

spm_coords_col(sspm_object)

\S4method{spm_coords_col}{sspm}(sspm_object)

\S4method{spm_coords_col}{sspm_discrete}(sspm_object)

spm_coords_col(object) <- value

\S4method{spm_coords_col}{sspm}(object) <- value

spm_time_col(sspm_object)

\S4method{spm_time_col}{sspm}(sspm_object)

\S4method{spm_time_col}{sspm_discrete}(sspm_object)

spm_time_col(object) <- value

\S4method{spm_time_col}{sspm}(object) <- value

spm_boundaries(sspm_object)

\S4method{spm_boundaries}{sspm}(sspm_object)

spm_boundaries(object) <- value

\S4method{spm_boundaries}{sspm_discrete}(object) <- value

\S4method{spm_boundaries}{sspm}(object) <- value

spm_discret_method(sspm_object)

\S4method{spm_discret_method}{sspm}(sspm_object)

\S4method{spm_discret_method}{sspm_discrete}(sspm_object)

spm_discret_method(object) <- value

\S4method{spm_discret_method}{sspm_discrete}(object) <- value

\S4method{spm_boundaries}{sspm}(object) <- value

spm_patches(sspm_object)

\S4method{spm_patches}{sspm}(sspm_object)

\S4method{spm_patches}{sspm_discrete}(sspm_object)

spm_patches(object) <- value

\S4method{spm_patches}{sspm_discrete}(object) <- value

\S4method{spm_patches}{sspm}(object) <- value

spm_points(sspm_object)

\S4method{spm_points}{sspm}(sspm_object)

\S4method{spm_points}{sspm_discrete}(sspm_object)

spm_points(object) <- value

\S4method{spm_points}{sspm_discrete}(object) <- value

\S4method{spm_points}{sspm}(object) <- value

spm_mapped_datasets(sspm_object)

\S4method{spm_mapped_datasets}{sspm}(sspm_object)

\S4method{spm_mapped_datasets}{sspm_discrete}(sspm_object)

spm_mapped_datasets(object) <- value

\S4method{spm_mapped_datasets}{sspm_discrete}(object) <- value

\S4method{spm_mapped_datasets}{sspm}(object) <- value

spm_mapped_formulas(sspm_object)

\S4method{spm_mapped_formulas}{sspm}(sspm_object)

\S4method{spm_mapped_formulas}{sspm_discrete}(sspm_object)

spm_mapped_formulas(object) <- value

\S4method{spm_mapped_formulas}{sspm_discrete}(object) <- value

\S4method{spm_mapped_formulas}{sspm}(object) <- value
}
\arguments{
\item{sspm_object}{\strong{[sspm OR adjacent]} An object of class
\link[=sspm-class]{sspm} or others derivative classes.}

\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
All methods described here allow to access the elements of contained in
objects of class \link[=sspm-class]{sspm} and others derivative classes
(\code{sspm_discrete}, etc...).
}
\section{Functions}{
\itemize{
\item \code{spm_name,sspm-method}: TODO

\item \code{spm_name<-}: TODO

\item \code{spm_name<-,sspm-method}: TODO

\item \code{spm_base_dataset}: TODO

\item \code{spm_base_dataset,sspm-method}: TODO

\item \code{spm_base_dataset,sspm_discrete-method}: TODO

\item \code{spm_base_dataset<-}: TODO

\item \code{spm_base_dataset<-,sspm-method}: TODO

\item \code{spm_datasets}: TODO

\item \code{spm_datasets,sspm-method}: TODO

\item \code{spm_datasets,sspm_discrete-method}: TODO

\item \code{spm_datasets,sspm_data-method}: TODO

\item \code{spm_unique_ID}: TODO

\item \code{spm_unique_ID,sspm-method}: TODO

\item \code{spm_unique_ID,sspm_discrete-method}: TODO

\item \code{spm_unique_ID<-}: TODO

\item \code{spm_unique_ID<-,sspm-method}: TODO

\item \code{spm_coords_col}: TODO

\item \code{spm_coords_col,sspm-method}: TODO

\item \code{spm_coords_col,sspm_discrete-method}: TODO

\item \code{spm_coords_col<-}: TODO

\item \code{spm_coords_col<-,sspm-method}: TODO

\item \code{spm_time_col}: TODO

\item \code{spm_time_col,sspm-method}: TODO

\item \code{spm_time_col,sspm_discrete-method}: TODO

\item \code{spm_time_col<-}: TODO

\item \code{spm_time_col<-,sspm-method}: TODO

\item \code{spm_boundaries}: TODO

\item \code{spm_boundaries,sspm-method}: TODO

\item \code{spm_boundaries<-}: TODO

\item \code{spm_boundaries<-,sspm_discrete-method}: TODO

\item \code{spm_boundaries<-,sspm-method}: TODO

\item \code{spm_discret_method}: TODO

\item \code{spm_discret_method,sspm-method}: TODO

\item \code{spm_discret_method,sspm_discrete-method}: TODO

\item \code{spm_discret_method<-}: TODO

\item \code{spm_discret_method<-,sspm_discrete-method}: TODO

\item \code{spm_boundaries<-,sspm-method}: TODO

\item \code{spm_patches}: TODO

\item \code{spm_patches,sspm-method}: TODO

\item \code{spm_patches,sspm_discrete-method}: TODO

\item \code{spm_patches<-}: TODO

\item \code{spm_patches<-,sspm_discrete-method}: TODO

\item \code{spm_patches<-,sspm-method}: TODO

\item \code{spm_points}: TODO

\item \code{spm_points,sspm-method}: TODO

\item \code{spm_points,sspm_discrete-method}: TODO

\item \code{spm_points<-}: TODO

\item \code{spm_points<-,sspm_discrete-method}: TODO

\item \code{spm_points<-,sspm-method}: TODO

\item \code{spm_mapped_datasets}: TODO

\item \code{spm_mapped_datasets,sspm-method}: TODO

\item \code{spm_mapped_datasets,sspm_discrete-method}: TODO

\item \code{spm_mapped_datasets<-}: TODO

\item \code{spm_mapped_datasets<-,sspm_discrete-method}: TODO

\item \code{spm_mapped_datasets<-,sspm-method}: TODO

\item \code{spm_mapped_formulas}: TODO

\item \code{spm_mapped_formulas,sspm-method}: TODO

\item \code{spm_mapped_formulas,sspm_discrete-method}: TODO

\item \code{spm_mapped_formulas<-}: TODO

\item \code{spm_mapped_formulas<-,sspm_discrete-method}: TODO

\item \code{spm_mapped_formulas<-,sspm-method}: TODO
}}

