% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_S4_all_classes_definitions.R
\docType{class}
\name{sspm-class}
\alias{sspm-class}
\title{sspm model class}
\description{
The \strong{\code{sspm}} model object, made from biomass, predictor and catch data.
}
\section{Slots}{

\describe{
\item{\code{datasets}}{\strong{[list]} List of
\link[=sspm_dataset-class]{sspm_dataset} that define variables in the SPM model.}

\item{\code{time}}{\strong{[character]} The column of \code{data} that represents the
temporal dimension of the dataset.}

\item{\code{uniqueID}}{\strong{[character]} The column of \code{datasets} that is unique for
all rows of the data matrix.}

\item{\code{boundaries}}{\strong{[sf]} Spatial boundaries (polygons).}

\item{\code{smoothed_data}}{\strong{[ANY (sf)]} The smoothed data.}

\item{\code{smoothed_vars}}{\strong{[character]} A vector storing the smoothed vars.}

\item{\code{is_split}}{\strong{[logical]} Whether this object has been split into
train/test sets.}
}}

