% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspm.R
\name{sspm}
\alias{sspm}
\title{Create a \code{sspm} model object}
\usage{
sspm(model_name = "My sspm model", boundaries)
}
\arguments{
\item{model_name}{\strong{[character]} The name to be given to the model}

\item{boundaries}{\strong{[sf]} The spatial boundaries (polygons) for the model.}

\item{...}{TBD}
}
\value{
An object of class  \link[=sspm-class]{sspm}.
}
\description{
The starting point of the \code{sspm} workflow: creating a base \code{sspm} object.
}
