% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspm-package.R
\docType{package}
\name{sspm-package}
\alias{sspm-package}
\title{sspm: Spatial Surplus Production Model Framework for Northern Shrimp Populations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implement a GAM-based (Generalized Additive Models) spatial surplus production model (spatial SPM), aimed at modeling northern shrimp population in Atlantic Canada but potentially to any stock in any location. The package is opinionated in its implementation of SPMs as it internally makes the choice to use penalized spatial gams with time lags. However, it also aims to provide options for the user to customize their model. The methods are described in Pedersen et al. (2022, \url{https://www.dfo-mpo.gc.ca/csas-sccs/Publications/ResDocs-DocRech/2022/2022_062-eng.html}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedersen-fisheries-lab.github.io/sspm/}
  \item Report bugs at \url{https://github.com/pedersen-fisheries-lab/sspm/issues}
}

}
\author{
\strong{Maintainer}: Valentin Lucet \email{valentin.lucet@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Eric Pedersen \email{eric.pedersen@concordia.ca}
}

}
\keyword{internal}
