% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_S4_all_classes_definitions.R
\docType{class}
\name{sspm_fit-class}
\alias{sspm_fit-class}
\title{sspm fit}
\description{
The fit object for a sspm model
}
\section{Slots}{

\describe{
\item{\code{smoothed_data}}{\strong{[ANY (sf)]} The smoothed data.}

\item{\code{time}}{\strong{[character]} The column of \code{smoothed_data} that
represents the temporal dimension of the dataset.}

\item{\code{uniqueID}}{\strong{[character]} The column of \code{smoothed_data} that is unique
for all rows of the data matrix.}

\item{\code{formula}}{\strong{[list]} The \link[=sspm_formula-class]{sspm_formula} object that
specifies the spm model.}

\item{\code{boundaries}}{\strong{[sf]} Spatial boundaries (polygons).}

\item{\code{fit}}{\strong{[bam]} The fit of the spm model.}
}}

